@@ -1,105 +1,47 @@
 require 'csv'
+require 'string'
 
-class String
-  def humanized_all_first_capitals
-    self.humanize.split(' ').map{|w| w.capitalize}.join(' ')
-  end
-end
+class Organisation < BaseOrganisation
 
-class Organisation < ActiveRecord::Base
-  #validates_presence_of :website, :with => /http:\/\//
-  acts_as_paranoid
-  validates_url :website, :prefferred_scheme => 'http://', :if => Proc.new{|org| org.website.present?}
-  validates_url :donation_info, :prefferred_scheme => 'http://', :if => Proc.new{|org| org.donation_info.present?}
-
-  # http://stackoverflow.com/questions/10738537/lazy-geocoding
-  has_many :users
   has_many :volunteer_ops
-  has_many :category_organisations
-  has_many :categories, :through => :category_organisations
-  # Setup accessible (or protected) attributes for your model
-  # prevents mass assignment on other fields not in this list
-  #attr_accessible :name, :description, :address, :postcode, :email, :website, :telephone, :donation_info, :publish_address, :publish_phone, :publish_email, :category_organisations_attributes
-  accepts_nested_attributes_for :users
-  accepts_nested_attributes_for :category_organisations,
-                                :allow_destroy => true
-  scope :order_by_most_recent, -> {order('updated_at DESC')}
+  has_many :users
+  has_many :edits, class_name: 'ProposedOrganisationEdit', :dependent => :destroy
+
+  accepts_nested_attributes_for :users # TODO check if needed
+
+  scope :order_by_most_recent, -> { order('organisations.updated_at DESC') }
   scope :not_null_email, lambda {where("organisations.email <> ''")}
+
   # Should we not use :includes, which pulls in extra data? http://nlingutla.com/blog/2013/04/21/includes-vs-joins-in-rails/
   # Alternative => :joins('LEFT OUTER JOIN users ON users.organisation_id = organisations.id)
   # Difference between inner and outer joins: http://stackoverflow.com/a/38578/2197402
+
   scope :null_users, lambda { includes(:users).where("users.organisation_id IS NULL").references(:users) }
   scope :without_matching_user_emails, lambda {where("organisations.email NOT IN (#{User.select('email').to_sql})")}
 
   after_save :uninvite_users, if: ->{ email_changed? }
 
-  def not_updated_recently?
-    updated_at < 365.day.ago
-  end
-
   def uninvite_users
     users.invited_not_accepted.update_all(organisation_id: nil)
   end
 
-  # For the geocoder gem
-  geocoded_by :full_address
-  after_validation :geocode, if: -> { run_geocode? }
-
-  def run_geocode?
-    ## http://api.rubyonrails.org/classes/ActiveModel/Dirty.html
-    address_changed? or (address.present? and not_geocoded?)
-  end
-
-  def not_geocoded?
-    latitude.blank? and longitude.blank?
-  end
-
-  #TODO: Give this TLC and refactor the flow or refactor out responsibilities
-  # This method both adds new editors and/or updates attributes
-  def update_attributes_with_admin(params)
-    email = params[:admin_email_to_add]
-    params.delete :admin_email_to_add
-    if email.blank?
-      return self.update_attributes(params)   # explicitly call with return to return boolean instead of nil
-    end
-    #Transactions are protective blocks where SQL statements are only permanent if they can all succeed as one atomic action.
-    ActiveRecord::Base.transaction do
-      usr = User.find_by_email(email)
-      if usr.present?
-        self.users << usr
-        return self.update_attributes(params)
-      else
-        self.errors.add(:administrator_email, "The user email you entered,'#{email}', does not exist in the system")
-        raise ActiveRecord::Rollback    # is this necessary? Doesn't the transaction block rollback the change with `usr` if update_attributes fails?
-      end
-    end
+  def update_attributes_with_superadmin(params)
+    email = extract_email_from(params)
+    return unless email.blank? || can_add_or_invite_admin?(email)
+    self.update_attributes(params)
   end
 
   def self.search_by_keyword(keyword)
      keyword = "%#{keyword}%"
-     self.where(contains_description(keyword).or(contains_name(keyword)))
+     where(contains_description?(keyword).or(contains_name?(keyword)))
   end
 
-  def self.filter_by_category(category_id)
-    return all unless category_id.present?
-    self.joins(:categories).where(is_in_category(category_id)) #do we need to sanitize category_id?
+  def self.filter_by_categories(category_ids)
+    joins(:categories)
+      .where(category_id.in category_ids)                 # at this point, orgs in multiple categories show up as duplicates
+      .group(organisation_id)                             # so we exploit this
+      .having(organisation_id.count.eq category_ids.size) # and return the orgs with correct number of duplicates
   end
-
-  def gmaps4rails_marker_attrs
-    if not_updated_recently_or_has_no_owner?
-      ['https://maps.gstatic.com/intl/en_ALL/mapfiles/markers2/measle.png',
-        'data-id' => id,
-        class: 'measle']
-    else
-      ['http://mt.googleapis.com/vt/icon/name=icons/spotlight/spotlight-poi.png',
-        'data-id' => id,
-       class: 'marker']
-    end
-  end
-
-  def full_address
-     "#{self.address}, #{self.postcode}"
-   end
 
   #Edit this if CSV 'schema' changes
   #value is the name of a column in csv file
@@ -143,12 +85,13 @@ class Organisation < ActiveRecord::Base
     CreateOrganisationFromArray.create(Organisation, row, validate)
   end
 
-  def self.create_and_validate(attributes) 
-    create!(attributes)
+  def self.create_and_validate(attributes)
+    create!(attributes.select{|k,v| !v.nil?})
   end
 
   def self.import_addresses(filename, limit, validation = true)
     import(filename, limit, validation) do |row, validation|
+       sleep 0.1
        create_from_array(row, validation)
     end
   end
@@ -178,7 +121,7 @@ class Organisation < ActiveRecord::Base
   def self.add_email(row, validation)
     orgs = where("UPPER(name) LIKE ? ","%#{row[0].try(:upcase)}%")
     return "#{row[0]} was not found\n" unless orgs && orgs[0] && orgs[0].email.blank?
-    orgs[0].email = row[7]
+    orgs[0].email = row[7].to_s
     orgs[0].save
     return "#{row[0]} was found\n"
   end
@@ -190,31 +133,59 @@ class Organisation < ActiveRecord::Base
       end
     end
   end
-  
-  def not_updated_recently_or_has_no_owner?
-    self.users.empty? || not_updated_recently?
-  end
 
   private
 
+  def embellish_invite_error_and_add_to_model(email, error_msg)
+    error_msg = ("Error: Email is invalid" == error_msg) ? "The user email you entered,'#{email}', is invalid" : error_msg
+    self.errors.add(:superadministrator_email, error_msg)
+  end
+
+  def add_and_notify(usr)
+    self.users << usr
+    org_admin_email = [usr.email]
+    OrgAdminMailer.new_org_admin(self, org_admin_email).deliver_now
+  end
+
+  def can_add_or_invite_admin?(email)
+    return false if email.blank?
+    usr = User.find_by_email(email)
+    return add_and_notify(usr) if usr.present?
+    result = ::SingleInviteJob.new(self, email).invite_user
+    return true if result.invited_user?
+    embellish_invite_error_and_add_to_model(email,result.error)
+    false
+  end
+
+  def extract_email_from(params)
+    email = params[:superadmin_email_to_add]
+    params.delete :superadmin_email_to_add
+    email
+  end
+
   def self.table
-    self.arel_table
+    arel_table
+  end
+
+  def self.organisation_id
+    table[:id]
   end
 
   def self.category_table
     Category.arel_table
   end
 
-  def self.is_in_category(category_id)
-    category_table[:id].eq(category_id)
+  def self.category_id
+    category_table[:id]
   end
-  
-  def self.contains_description(key)
+
+  def self.contains_description?(key)
     table[:description].matches(key)
   end
 
-  def self.contains_name(key)
+  def self.contains_name?(key)
     table[:name].matches(key)
   end
+
 
 end
