@@ -2,8 +2,10 @@ require 'webmock/cucumber'
 require 'uri-handler'
 include ApplicationHelper
 
-Then(/^I travel "(.*?)" days into the future$/) do |days|
-  Delorean.time_travel_to "#{days} days from now"
+Then(/^I travel a year plus "(.*?)" days into the future$/) do |days|
+  Timecop.travel(
+    Time.current.advance(years: 1, days: days.to_i)
+  )
 end
 Then(/^I should see the "(.*?)" image linked to "(.*?)"$/) do |image_alt, link|
   within("a[href='#{link}']") do
@@ -19,33 +21,33 @@ Then /^I should see permission denied$/ do
   page.should have_content PERMISSION_DENIED
 end
 
-Then(/^the Admin menu has a valid (.*?) link$/) do |link|
-  within('#menuAdmin > ul.dropdown-menu') do
+Then(/^the SuperAdmin menu has a valid (.*?) link$/) do |link|
+  within('#menuSuperAdmin > ul.dropdown-menu') do
     find('a', text: link).should_not be_nil
     click_link link
     current_path.should eq paths(link.downcase)
   end
 end
 
-Then /^"(.*?)" should be a charity admin for "(.*?)" charity$/ do |email, org|
+Then /^"(.*?)" should be a charity superadmin for "(.*?)" charity$/ do |email, org|
   org = Organisation.find_by_name org
   usr = User.find_by_email(email)
   expect(usr).not_to be_nil
   expect(org.users).to include usr
 end
 
-Then /^I should see the cannot add non registered user "(.*?)" as charity admin message$/ do |email|
+Then /^I should see the cannot add non registered user "(.*?)" as charity superadmin message$/ do |email|
   page.should have_content "The user email you entered,'#{email}', does not exist in the system"
 end
 
-And /^I add "(.*?)" as an admin for "(.*?)" charity$/ do |admin_email, charity|
+And /^I add "(.*?)" as a superadmin for "(.*?)" charity$/ do |superadmin_email, charity|
   steps %Q{ And I visit the edit page for the organisation named "#{charity}"}
-  fill_in 'organisation_admin_email_to_add', :with => admin_email
+  fill_in 'organisation_superadmin_email_to_add', :with => superadmin_email
   steps %Q{
   And I press "Update Organisation"}
 end
 
-Then /^I should see the no charity admins message$/ do
+Then /^I should see the no charity superadmins message$/ do
   expect(page).to have_content "This organisation has no admins yet"
 end
 
@@ -66,7 +68,7 @@ Then /^I should( not)? see an edit button for "(.*?)" volunteer opportunity$/ do
   expect(page).send(expectation_method, have_link('Edit', :href => edit_volunteer_op_path(op.id)))
 end
 
-Then /^I should( not)? see "(.*?)" in the charity admin email$/ do |negate,email|
+Then /^I should( not)? see "(.*?)" in the charity superadmin email$/ do |negate,email|
   expectation_method = negate ? :not_to : :to
   expect(page).send(expectation_method, have_selector('li', :text => email))
 end
@@ -81,19 +83,67 @@ When /^I search for "(.*?)"$/ do |text|
 end
 
 Given (/^I fill in the new charity page validly$/) do
-  stub_request_with_address("64 pinner road")
   fill_in 'organisation_address', :with => '64 pinner road'
   fill_in 'organisation_name', :with => 'Friendly charity'
+  fill_in 'organisation_postcode', :with => 'HA1 4HZ'
 end
+
 Given (/^I fill in the new charity page validly including the categories:$/) do |categories_table|
+  fill_in 'organisation_address', :with => '64 pinner road'
+  fill_in 'organisation_name', :with => 'Friendly charity'
+  fill_in 'organisation_postcode', :with => 'HA1 4HZ'
   categories_table.hashes.each do |cat|
     steps %Q{
       And I check the category "#{cat[:name]}"
     }
   end
-  stub_request_with_address("64 pinner road")
-  fill_in 'organisation_address', :with => '64 pinner road'
-  fill_in 'organisation_name', :with => 'Friendly charity'
+end
+
+Given(/^I am proposing an organisation$/) do
+  visit new_proposed_organisation_path
+  steps %Q{And I fill in the proposed charity page validly}
+end
+
+Given (/^I fill in the proposed charity page validly$/) do
+  proposed_org_fields.each do |key, val|
+    fill_in "proposed_organisation_#{key}", with: val
+  end
+  proposed_org_categories.each do |cat|
+    steps %Q{
+      And I check the category "#{cat}"
+    }
+  end
+end
+
+When(/^I check the confirmation box for "(.*?)"$/) do |text|
+  find(:xpath,"//label[text()='#{text}']/preceding-sibling::input[1]").set(true)
+end
+
+When(/^I uncheck the confirmation box for "(.*?)"$/) do |text|
+  find(:xpath,"//label[text()='#{text}']/preceding-sibling::input[1]").set(false)
+end
+
+Then(/^I should be on the proposed organisations show page for the organisation$/) do
+  expect(current_path).to eq proposed_organisation_path(ProposedOrganisation.find_by(name: proposed_org_fields[:name]))
+end
+
+Then(/^the proposed organisation should have been created$/) do
+  expect(ProposedOrganisation.find_by(name: 'Friendly charity')).not_to be_nil
+end
+
+Then (/the confirmation box named (.*) should be (checked|unchecked)$/) do |category, status|
+  assertion = (status == 'checked') ? :should : :should_not
+  page.find(:xpath, "//label[text()='#{category}']/preceding-sibling::input[1]").send(assertion, be_checked)
+end
+
+
+Then(/^I should see all the proposed organisation fields$/) do
+  proposed_org_fields.each_value do |value|
+    expect(page).to have_content(value)
+  end
+  proposed_org_categories.each do |cat|
+    expect(page).to have_content(cat)
+  end
 end
 
 Then /^the contact information should be available$/ do
@@ -102,12 +152,14 @@ Then /^the contact information should be available$/ do
     Then I should see "Contact Info Email us: contact@voluntaryactionharrow.org.uk Phone Us: 020 8861 5894 Write to Us: The Lodge, 64 Pinner Road, Harrow, Middlesex, HA1 4HZ Find Us: On Social Media (Click Here)"
    }
 end
+
 Then /^the about us should be available$/ do
   steps %Q{
     When I follow "About Us"
     Then I should see "About Us Supporting groups in Harrow We are a not-for-profit workers co-operative who support people and not-for-profit organisations to make a difference in their local community by: Working with local people and groups to identify local needs and develop appropriate action. Providing a range of services that help organisations to succeed. Supporting and encouraging the growth of co-operative movement. How do we support? Find out here (VAH in a nutshell) What is a Workers Co-operative? A workers co-operative is a business owned and democratically controlled by their employee members using co-operative principles. They are an attractive and increasingly relevant alternative to traditional investor owned models of enterprise. (Click here for more details)"
   }
 end
+
 Given /^I update "(.*?)" charity address to be "(.*?)"( when Google is indisposed)?$/ do |name, address, indisposed|
   steps %Q{
     Given I visit the show page for the organisation named "#{name}"
@@ -161,7 +213,6 @@ Given /^I edit the charity email to be "(.*?)"$/ do |email|
 end
 
 When /^I edit the charity address to be "(.*?)"$/ do |address|
-  stub_request_with_address(address)
   fill_in('organisation_address', :with => address)
 end
 
@@ -268,9 +319,19 @@ Then /^I should see "([^"]*)" and "([^"]*)"$/ do |text1, text2|
   expect(page).to have_content text2
 end
 
-Then /^I should( not)? see "((?:(?!before|").)+)"$/ do |negate, text|
+Then /^I should( not)? see "([^"]*)"$/ do |negate, text|
   expectation_method = negate ? :not_to : :to
   expect(page).send(expectation_method, have_content(text))
+end
+
+Then(/^I should see "(.*?)" within "(.*?)"$/) do |text, selector|
+  within('#' + selector) { expect(page).to have_content text}
+end
+
+Then(/^I should see the following:$/) do |table|
+  table.rows.each do |text|
+    expect(page).to have_content text.first
+  end
 end
 
 Then(/^I should( not)? see a link or button "(.*?)"$/) do |negate, link|
@@ -283,12 +344,21 @@ Then(/^the navbar should( not)? have a link to (.*?)$/) do |negate, link|
   within('#navbar') { expect(page).send(expectation_method, have_selector(:link_or_button, link)) }
 end
 
+Then(/^I should not see "(.*?)"  within "(.*?)"$/) do |text, selector|
+  within('.' + selector) { expect(page).not_to have_content text}
+end
+
+Given /^I update "(.*?)" charity postcode to be "(.*?)"$/ do |name, postcode|
+  steps %Q{And I visit the edit page for the organisation named "#{name}"}
+  fill_in('organisation_postcode', :with => postcode)
+  click_button 'Update Organisation'
+end
+
 Given /^I edit the charity address to be "(.*?)" when Google is indisposed$/ do |address|
   body = %Q({
 "results" : [],
 "status" : "OVER_QUERY_LIMIT"
 })
-  stub_request_with_address(address, body)
   fill_in('organisation_address', :with => address)
 end
 
@@ -298,6 +368,10 @@ end
 
 Then /^the address for "(.*?)" should be "(.*?)"$/ do |name, address|
   Organisation.find_by_name(name).address.should == address
+end
+
+Then /^the postcode for "(.*?)" should be "(.*?)"$/ do |name, postcode|
+  Organisation.find_by_name(name).postcode.should == postcode
 end
 
 When /^I fill in "(.*?)" with "(.*?)" within the navbar$/ do |field, value|
@@ -328,11 +402,10 @@ And /^I click "(.*)" on the "(.*)" page and stay there$/  do |link, org_name|
   }
 end
 
-
 Given /^"(.*)"'s request status for "(.*)" should be updated appropriately$/ do |email, org_name|
   steps %Q{
       And "#{email}"'s request for "#{org_name}" should be persisted
-      And I should see "You have requested admin status for #{Organisation.find_by_name(org_name).name}"
+      And I should see "You have requested superadmin status for #{Organisation.find_by_name(org_name).name}"
       And I should not see a link or button "This is my organisation"
     }
 end
@@ -375,7 +448,7 @@ When /^I approve "(.*?)"$/ do |email|
   end
 end
 
-Then(/^"(.*?)" is a charity admin of "(.*?)"$/) do |user_email, org_name|
+Then(/^"(.*?)" is an organisation admin of "(.*?)"$/) do |user_email, org_name|
   user = User.find_by_email(user_email)
   org = Organisation.find_by_name(org_name)
   user.organisation.should == org
@@ -427,6 +500,15 @@ Then /^the index should( not)? contain:$/ do |negative, table|
   end
 end
 
+Then /^the organisation results should( not)? contain:$/ do |negative, table|
+  expectation = negative ? :should_not : :should
+  table.raw.flatten.each do |cell|
+    within('#orgs_scroll') do
+      page.send(expectation, have_text(cell))
+    end
+  end
+end
+
 Then(/^I should see "([^"]*)" page before "([^"]*)"$/) do |first_item, second_item|
   page.body.should =~ /#{first_item}.*#{second_item}/m
 end
@@ -452,10 +534,10 @@ Given /^debugger$/ do
   debugger
   puts ""
 end
+
 Given /^I run the invite migration$/ do
 
 end
-
 
 Given(/^I can run the rake task "(.*?)"$/) do |task|
   stdout, stderr, status = Open3.capture3("#{task}")
