@@ -1,20 +1,45 @@
-def stub_request_with_address(address, body = nil)
-  filename = "#{address.gsub(/\s/, '_')}.json"
-  filename = File.read "test/fixtures/#{filename}"
-  stub_request(:any, /maps\.googleapis\.com/).
-      to_return(status => 200, :body => body || filename, :headers => {})
+Given(/^the following proposed organisations exist:$/) do |table|
+  require 'boolean'
+  table.hashes.each do |hash|
+    create_hash = {}
+    proposer = nil
+    hash.each_pair do |field_name, field_value|
+      if field_name != "proposer_email"
+        key_value_to_add = {field_name.to_sym => field_value}
+      else
+        proposer = User.find_by(email: field_value)
+      end
+      create_hash.merge! key_value_to_add unless key_value_to_add.nil?
+    end
+    proposed_org = ProposedOrganisation.new create_hash
+    proposed_org.users << proposer if proposer
+    proposed_org.save!
+  end
+end
+
+Given(/^a proposed organisation has been proposed by "(.*)"$/) do |user_email|
+  usr = User.find_by(email: user_email)
+  unsaved_proposed_organisation(usr).save!
+end
+
+Then(/^I should see the details of the proposed organisation$/) do
+  [:name, :description, :email, :address, :postcode].each do |key|
+    expect(page).to have_content unsaved_proposed_organisation[key]
+  end
+  ["Donate to Friendly Charity now!", 'We are a not for profit organisation registered or working in Harrow'].each do |value|
+      expect(page).to have_content value
+  end
 end
 
 Given /^the following organisations exist:$/ do |organisations_table|
   organisations_table.hashes.each do |org|
-    stub_request_with_address(org['address'])
     Organisation.create! org
   end
 end
 
 Given /^the following users are registered:$/ do |users_table|
   users_table.hashes.each do |user|
-    user["admin"] = user["admin"] == "true"
+    user["superadmin"] = user["superadmin"] == "true"
     user["organisation"] = Organisation.find_by_name(user["organisation"])
     user["pending_organisation"] = Organisation.find_by_name(user["pending_organisation"])
     User.create! user
@@ -38,6 +63,14 @@ Given /^the following categories_organisations exist:$/ do |join_table|
   join_table.hashes.each do |row|
      cat = Category.find_by_name row[:category]
      org = Organisation.find_by_name row[:organisation]
+     org.categories << cat
+  end
+end
+
+Given /^the proposed organisations have the categories:$/ do |join_table|
+  join_table.hashes.each do |row|
+     cat = Category.find_by_name row[:category]
+     org = ProposedOrganisation.find_by_name row[:proposed_organisation]
      org.categories << cat
   end
 end
@@ -69,4 +102,11 @@ end
 Then(/^the organisation "([^"]*)" should be deleted$/) do |name|
   org = Organisation.only_deleted.find_by_name name
   expect(org).not_to be_nil
+end
+
+Then(/^the "(.*?) proposed edits for the organisation named "(.*?)" should only be soft deleted$/) do |number, name|
+  number = number.to_i
+  org = Organisation.with_deleted.find_by(name: name)
+  expect(ProposedOrganisationEdit.where(organisation: org)).to be_empty
+  expect(ProposedOrganisationEdit.with_deleted.where(organisation: org).size).to eq number
 end
