@@ -36,7 +36,7 @@ describe ApplicationController, :type => :controller, :helpers => :controllers d
   end
 
   describe '#after_sign_in_path_for' do
-    let(:user) {make_current_user_nonadmin}
+    let(:user) {make_current_user_nonsuperadmin}
     context 'user not associated with any org' do
       it 'should redirect to root' do
         allow(user).to receive_messages :organisation => nil, :pending_organisation_id => nil
@@ -46,14 +46,14 @@ describe ApplicationController, :type => :controller, :helpers => :controllers d
 
     context 'user is org owner no previous url' do
       it 'should redirect to root' do
-        allow(user).to receive_messages :organisation => mock_model(Organisation, id: 1, not_updated_recently?: true)
+        allow(user).to receive_messages :organisation => mock_model(Organisation, id: 1, has_been_updated_recently?: false)
         expect(controller.after_sign_in_path_for(user)).to eq '/organisations/1'
       end
     end
 
     context 'user is org owner with previous url' do
       it 'should redirect to root' do
-        allow(user).to receive_messages :organisation => mock_model(Organisation, id: 1, not_updated_recently?: true)
+        allow(user).to receive_messages :organisation => mock_model(Organisation, id: 1, has_been_updated_recently?: false)
         session[:previous_url] = 'i/was/here'
         expect(controller.after_sign_in_path_for(user)).to eq '/organisations/1'
       end
@@ -61,7 +61,7 @@ describe ApplicationController, :type => :controller, :helpers => :controllers d
   end
 
   it '#after_accept_path_for' do
-    user = make_current_user_nonadmin
+    user = make_current_user_nonsuperadmin
     allow(user).to receive_messages :organisation => nil
 
     expect(controller.after_accept_path_for(user)).to eq '/'
@@ -114,10 +114,10 @@ describe ApplicationController, :type => :controller, :helpers => :controllers d
 
     context '#authorize' do
       it 'Unauthorized: redirects to root_path and displays flash' do
-        allow(controller).to receive_messages admin?: false
+        allow(controller).to receive_messages superadmin?: false
         expect(controller).to receive(:redirect_to).with(root_path) { true } # calling original raises errors
         expect(controller.flash).to receive(:[]=)
-          .with(:error, 'You must be signed in as an admin to perform this action!')
+          .with(:error, 'You must be signed in as a superadmin to perform this action!')
           .and_call_original
         expect(controller.instance_eval { authorize }).to be false
         # can't assert `redirect_to root_path`: http://owowthathurts.blogspot.com/2013/08/rspec-response-delegation-error-fix.html
@@ -125,22 +125,22 @@ describe ApplicationController, :type => :controller, :helpers => :controllers d
       end
 
       it 'Authorized: allows execution to continue' do
-        allow(controller).to receive_messages admin?: true
+        allow(controller).to receive_messages superadmin?: true
         expect(controller.instance_eval { authorize }).to be nil
       end
     end
 
-    context '#admin?' do
+    context '#superadmin?' do
       it 'returns nil when current_user is nil' do
         allow(controller).to receive_messages current_user: nil
-        expect(controller.instance_eval { admin? }).to be_nil
+        expect(controller.instance_eval { superadmin? }).to be_nil
       end
 
-      it 'otherwise depends on { current_user.admin? }' do
-        expect(user).to receive(:admin?) { false }
-        expect(controller.instance_eval { admin? }).to be false
-        expect(user).to receive(:admin?) { true }
-        expect(controller.instance_eval { admin? }).to be true
+      it 'otherwise depends on { current_user.superadmin? }' do
+        expect(user).to receive(:superadmin?) { false }
+        expect(controller.instance_eval { superadmin? }).to be false
+        expect(user).to receive(:superadmin?) { true }
+        expect(controller.instance_eval { superadmin? }).to be true
       end
     end
   end
