@@ -3,12 +3,6 @@ require 'rails_helper'
 describe OrganisationsController, :type => :controller do
   let(:category_html_options) { [['cat1', 1], ['cat2', 2]] }
 
-  # shouldn't this be done in spec_helper.rb?
-  before :suite do
-    FactoryGirl.factories.clear
-    FactoryGirl.find_definitions
-  end
-
   # http://stackoverflow.com/questions/10442159/rspec-as-null-object
   # doesn't calling as_null_object on a mock negate the need to stub anything?
   def double_organisation(stubs={})
@@ -21,8 +15,9 @@ describe OrganisationsController, :type => :controller do
 
   describe "#build_map_markers" do
     render_views
-    let(:org) { create :organisation }
-    subject { JSON.parse(controller.send(:build_map_markers, org)).first }
+    let!(:org) { create :organisation }
+    let(:org_relation){Organisation.all}
+    subject { JSON.parse(controller.send(:build_map_markers, org_relation)).first }
     it { expect(subject['lat']).to eq org.latitude }
     it { expect(subject['lng']).to eq org.longitude }
     it { expect(subject['infowindow']).to include org.id.to_s }
@@ -30,11 +25,19 @@ describe OrganisationsController, :type => :controller do
     it { expect(subject['infowindow']).to include org.description }
     context 'markers without coords omitted' do
       let!(:org) { create :organisation, address: '150 pinner rd', latitude: nil, longitude: nil }
-      it { expect(JSON.parse(controller.send(:build_map_markers, org))).to be_empty }
+      it { expect(JSON.parse(controller.send(:build_map_markers, org_relation))).to be_empty }
     end
   end
 
   describe "GET search" do
+
+    let(:category_params) do
+      {
+        "what_id" => "",
+        "who_id"  => "",
+        "how_id"  => "",
+      }
+    end
 
     context 'setting appropriate view vars for all combinations of input' do
       let!(:double_organisation) { create :organisation }
@@ -43,61 +46,52 @@ describe OrganisationsController, :type => :controller do
       let(:category) { double('Category') }
       before(:each) do
         expect(controller).to receive(:build_map_markers).and_return(markers)
-        expect(Category).to receive(:html_drop_down_options).and_return(category_html_options)
+        allow(Category).to receive(:what_they_do) { Category.all }
+        allow(Category).to receive(:who_they_help) { Category.all }
+        allow(Category).to receive(:how_they_help) { Category.all }
       end
 
       it "orders search results by most recent" do
         expect(Organisation).to receive(:order_by_most_recent).and_return(result)
-        allow(result).to receive_message_chain(:search_by_keyword, :filter_by_category).with('test').with(nil).and_return(result)
-        get :search, :q => 'test'
+        expect(result).to receive(:search_by_keyword).with('test').and_return(result)
+        get :search, { q: 'test' }.merge(category_params)
         expect(assigns(:organisations)).to eq([double_organisation])
       end
 
       it "sets up appropriate values for view vars: query_term, organisations and markers" do
         expect(Organisation).to receive(:search_by_keyword).with('test').and_return(result)
-        expect(result).to receive(:filter_by_category).with('1').and_return(result)
-        expect(Category).to receive(:find_by_id).with("1").and_return(category)
-        get :search, :q => 'test', "category" => {"id" => "1"}
-        expect(assigns(:query_term)).to eq 'test'
-        expect(assigns(:category)).to eq category
+        expect(result).to receive(:filter_by_categories).with(['1']).and_return(result)
+        get :search, { q: 'test' }.merge(category_params).merge("what_id" => "1")
+        expect(assigns(:parsed_params).query_term).to eq 'test'
       end
 
       it "handles lack of category gracefully" do
         expect(Organisation).to receive(:search_by_keyword).with('test').and_return(result)
-        expect(result).to receive(:filter_by_category).with(nil).and_return(result)
-        get :search, :q => 'test'
-        expect(assigns(:query_term)).to eq 'test'
+        get :search, { q: 'test' }.merge(category_params)
+        expect(assigns(:parsed_params).query_term).to eq 'test'
       end
 
       it "handles lack of query term gracefully" do
-        expect(Organisation).to receive(:search_by_keyword).with(nil).and_return(result)
-        expect(result).to receive(:filter_by_category).with(nil).and_return(result)
-        get :search
-        expect(assigns(:query_term)).to eq nil
+        expect(Organisation).to receive(:search_by_keyword).with("").and_return(result)
+        get :search, category_params.merge({q: ''})
+        expect(assigns(:parsed_params).query_term).to eq("")
       end
 
       it "handles lack of id gracefully" do
         expect(Organisation).to receive(:search_by_keyword).with('test').and_return(result)
-        expect(result).to receive(:filter_by_category).with(nil).and_return(result)
-        get :search, :q => 'test', "category" => nil
-        expect(assigns(:query_term)).to eq 'test'
-      end
-
-      it "handles empty string id gracefully" do
-        expect(Organisation).to receive(:search_by_keyword).with('test').and_return(result)
-        expect(result).to receive(:filter_by_category).with("").and_return(result)
-        get :search, :q => 'test', "category" => {"id" => ""}
-        expect(assigns(:query_term)).to eq 'test'
+        get :search, { q: 'test' }.merge(category_params)
+        expect(assigns(:parsed_params).query_term).to eq 'test'
       end
 
       after(:each) do
         expect(response).to render_template 'index'
-        expect(response).to render_template 'layouts/two_columns'
+        expect(response).to render_template 'layouts/two_columns_with_map'
         expect(assigns(:organisations)).to eq([double_organisation])
         expect(assigns(:markers)).to eq(markers)
-        expect(assigns(:category_options)).to eq category_html_options
+        expect(assigns(:cat_name_ids)).to eq({what: [], how: [], who: []})
       end
     end
+
     # TODO figure out how to make this less messy
     it "assigns to flash.now but not flash when search returns no results" do
       expect(controller).to receive(:build_map_markers).and_return('my markers')
@@ -105,13 +99,12 @@ describe OrganisationsController, :type => :controller do
       result = Organisation.all
       expect(result).to receive(:empty?).and_return(true)
       expect(Organisation).to receive(:search_by_keyword).with('no results').and_return(result)
-      expect(result).to receive(:filter_by_category).with('1').and_return(result)
+      expect(result).to receive(:filter_by_categories).with(['1']).and_return(result)
       category = double('Category')
-      expect(Category).to receive(:find_by_id).with("1").and_return(category)
       expect_any_instance_of(ActionDispatch::Flash::FlashHash).to receive(:now).and_return double_now_flash
       expect_any_instance_of(ActionDispatch::Flash::FlashHash).not_to receive(:[]=)
       expect(double_now_flash).to receive(:[]=).with(:alert, SEARCH_NOT_FOUND)
-      get :search, :q => 'no results', "category" => {"id" => "1"}
+      get :search, { q: 'no results' }.merge(category_params).merge("what_id" => "1")
     end
 
     it "does not set up flash nor flash.now when search returns results" do
@@ -120,10 +113,9 @@ describe OrganisationsController, :type => :controller do
       expect(controller).to receive(:build_map_markers).and_return(markers)
       expect(result).to receive(:empty?).and_return(false)
       expect(Organisation).to receive(:search_by_keyword).with('some results').and_return(result)
-      expect(result).to receive(:filter_by_category).with('1').and_return(result)
+      expect(result).to receive(:filter_by_categories).with(['1']).and_return(result)
       category = double('Category')
-      expect(Category).to receive(:find_by_id).with("1").and_return(category)
-      get :search, :q => 'some results', "category" => {"id" => "1"}
+      get :search, { q: 'some results' }.merge(category_params).merge("what_id" => "1")
       expect(flash.now[:alert]).to be_nil
       expect(flash[:alert]).to be_nil
     end
@@ -134,13 +126,11 @@ describe OrganisationsController, :type => :controller do
       result = Organisation.all
       markers='my markers'
       expect(controller).to receive(:build_map_markers).and_return(markers)
-      expect(Category).to receive(:html_drop_down_options).and_return(category_html_options)
       expect(Organisation).to receive(:order_by_most_recent).and_return(result)
-      allow(result).to receive_message_chain(:page, :per).and_return(result)
       get :index
       expect(assigns(:organisations)).to eq(result)
       expect(assigns(:markers)).to eq(markers)
-      expect(response).to render_template 'layouts/two_columns'
+      expect(response).to render_template 'layouts/two_columns_with_map'
     end
   end
 
@@ -148,8 +138,7 @@ describe OrganisationsController, :type => :controller do
     let(:real_org){create(:organisation)}
     before(:each) do
       @user = double("User")
-      allow(@user).to receive(:pending_admin?)
-      allow(Organisation).to receive(:find).with('37') { real_org}
+      allow(@user).to receive(:pending_org_admin?)
       allow(@user).to receive(:can_edit?)
       allow(@user).to receive(:can_delete?)
       allow(@user).to receive(:can_create_volunteer_ops?)
@@ -157,59 +146,58 @@ describe OrganisationsController, :type => :controller do
       allow(controller).to receive(:current_user).and_return(@user)
     end
 
-    it 'should use a two_column layout' do
-      get :show, :id => '37'
-      expect(response).to render_template 'layouts/two_columns'
+    it 'should use a two_column with map layout' do
+      get :show, :id => real_org.id.to_s
+      expect(response).to render_template 'layouts/two_columns_with_map'
     end
 
     it "assigns the requested organisation as @organisation and appropriate markers" do
       markers='my markers'
       @org = real_org
       expect(controller).to receive(:build_map_markers).and_return(markers)
-      expect(Organisation).to receive(:find).with('37') { @org }
-      get :show, :id => '37'
-      expect(assigns(:organisation)).to be(real_org)
+      get :show, :id => real_org.id.to_s
+      expect(assigns(:organisation)).to eq(real_org)
       expect(assigns(:markers)).to eq(markers)
     end
 
     context "editable flag is assigned to match user permission" do
       it "user with permission leads to editable flag true" do
         expect(@user).to receive(:can_edit?).with(real_org).and_return(true)
-        get :show, :id => 37
+        get :show, id: real_org.id.to_s
         expect(assigns(:editable)).to be(true)
       end
 
       it "user without permission leads to editable flag false" do
         expect(@user).to receive(:can_edit?).with(real_org).and_return(true)
-        get :show, :id => 37
+        get :show, id: real_org.id.to_s
         expect(assigns(:editable)).to be(true)
       end
 
       it 'when not signed in editable flag is nil' do
         allow(controller).to receive(:current_user).and_return(nil)
-        get :show, :id => 37
+        get :show, id: real_org.id.to_s
         expect(assigns(:editable)).to be_nil
       end
     end
 
     context "grabbable flag is assigned to match user permission" do
-      it 'assigns grabbable to true when user can request org admin status' do
+      it 'assigns grabbable to true when user can request org superadmin status' do
         allow(@user).to receive(:can_edit?)
         expect(@user).to receive(:can_request_org_admin?).with(real_org).and_return(true)
         allow(controller).to receive(:current_user).and_return(@user)
-        get :show, :id => 37
+        get :show, :id => real_org.id.to_s
         expect(assigns(:grabbable)).to be(true)
       end
-      it 'assigns grabbable to false when user cannot request org admin status' do
+      it 'assigns grabbable to false when user cannot request org superadmin status' do
         allow(@user).to receive(:can_edit?)
         expect(@user).to receive(:can_request_org_admin?).with(real_org).and_return(false)
         allow(controller).to receive(:current_user).and_return(@user)
-        get :show, :id => 37
+        get :show, :id => real_org.id.to_s
         expect(assigns(:grabbable)).to be(false)
       end
       it 'when not signed in grabbable flag is true' do
         allow(controller).to receive(:current_user).and_return(nil)
-        get :show, :id => 37
+        get :show, :id => real_org.id.to_s
         expect(assigns(:grabbable)).to be true
       end
     end
@@ -218,13 +206,13 @@ describe OrganisationsController, :type => :controller do
       context 'depends on can_create_volunteer_ops?' do
         it 'true' do
           expect(@user).to receive(:can_create_volunteer_ops?) { true }
-          get :show, :id => 37
+          get :show, :id => real_org.id.to_s
           expect(assigns(:can_create_volunteer_op)).to be true
         end
 
         it 'false' do
           expect(@user).to receive(:can_create_volunteer_ops?) { false }
-          get :show, :id => 37
+          get :show, :id => real_org.id.to_s
           expect(assigns(:can_create_volunteer_op)).to be false
         end
       end
@@ -232,7 +220,7 @@ describe OrganisationsController, :type => :controller do
       it 'will not be called when current user is nil' do
         allow(controller).to receive_messages current_user: nil
         expect(@user).not_to receive :can_create_volunteer_ops?
-        get :show, :id => 37
+        get :show, :id => real_org.id.to_s
         expect(assigns(:can_create_volunteer_op)).to be nil
       end
     end
@@ -252,9 +240,9 @@ describe OrganisationsController, :type => :controller do
         expect(assigns(:organisation)).to be(double_organisation)
       end
 
-      it 'should use a two_column layout' do
+      it 'should use a two_column with map layout' do
         get :new
-        expect(response).to render_template 'layouts/two_columns'
+        expect(response).to render_template 'layouts/two_columns_with_map'
       end
     end
 
@@ -267,7 +255,7 @@ describe OrganisationsController, :type => :controller do
   end
 
   describe "GET edit" do
-    let(:real_org) { create(:organisation) }
+    let(:org) { create(:organisation) }
 
     context "while signed in as user who can edit" do
       before(:each) do
@@ -278,10 +266,9 @@ describe OrganisationsController, :type => :controller do
       end
 
       it "assigns the requested organisation as @organisation" do
-        allow(Organisation).to receive(:find).with('37') { real_org }
-        get :edit, :id => '37'
-        expect(assigns(:organisation)).to be(real_org)
-        expect(response).to render_template 'layouts/two_columns'
+        get :edit, id: org.id
+        expect(assigns(:organisation)).to eq org
+        expect(response).to render_template 'layouts/two_columns_with_map'
       end
     end
     context "while signed in as user who cannot edit" do
@@ -293,9 +280,8 @@ describe OrganisationsController, :type => :controller do
       end
 
       it "redirects to organisation view" do
-        allow(Organisation).to receive(:find).with('37') { real_org }
-        get :edit, :id => '37'
-        expect(response).to redirect_to organisation_url(37)
+        get :edit, id: org.id
+        expect(response).to redirect_to organisation_url(org)
       end
     end
     #TODO: way to dry out these redirect specs?
@@ -308,10 +294,10 @@ describe OrganisationsController, :type => :controller do
   end
 
   describe "POST create", :helpers => :controllers do
-    context "while signed in as admin" do
+    context "while signed in as superadmin" do
       let!(:org) { build :organisation }
       before(:each) do
-        expect(make_current_user_admin).to receive(:admin?).and_return true
+        expect(make_current_user_superadmin).to receive(:superadmin?).and_return true
       end
 
       describe "with valid params" do
@@ -329,7 +315,7 @@ describe OrganisationsController, :type => :controller do
       end
 
       describe "with invalid params" do
-        after(:each) { expect(response).to render_template 'layouts/two_columns' }
+        after(:each) { expect(response).to render_template 'two_columns_with_map' }
 
         it "assigns a newly created but unsaved organisation as @organisation" do
           allow(Organisation).to receive(:new){ double_organisation(:save => false) }
@@ -353,9 +339,9 @@ describe OrganisationsController, :type => :controller do
       end
     end
 
-    context "while signed in as non-admin" do
+    context "while signed in as non-superadmin" do
       before(:each) do
-        expect(make_current_user_nonadmin).to receive(:admin?).and_return(false)
+        expect(make_current_user_nonsuperadmin).to receive(:superadmin?).and_return(false)
       end
 
       describe "with valid params" do
@@ -395,7 +381,7 @@ describe OrganisationsController, :type => :controller do
       describe "with valid params" do
         it "updates org for e.g. donation_info url" do
           expect(Organisation).to receive(:find).with('37') { org }
-          expect(org).to receive(:update_attributes_with_admin).with({'donation_info' => 'http://www.friendly.com/donate', 'admin_email_to_add' => nil})
+          expect(org).to receive(:update_attributes_with_superadmin).with({'donation_info' => 'http://www.friendly.com/donate', 'superadmin_email_to_add' => nil})
           put :update, :id => '37', :organisation => {'donation_info' => 'http://www.friendly.com/donate'}
         end
 
@@ -413,16 +399,16 @@ describe OrganisationsController, :type => :controller do
       end
 
       describe "with invalid params" do
-        after(:each) { expect(response).to render_template 'layouts/two_columns' }
+        after(:each) { expect(response).to render_template 'layouts/two_columns_with_map' }
 
         it "assigns the organisation as @organisation" do
-          allow(Organisation).to receive(:find) { double_organisation(:update_attributes_with_admin => false) }
+          allow(Organisation).to receive(:find) { double_organisation(:update_attributes_with_superadmin => false) }
           put :update, :id => "1", :organisation => {'these' => 'params'}
           expect(assigns(:organisation)).to be(double_organisation)
         end
 
         it "re-renders the 'edit' template" do
-          allow(Organisation).to receive(:find) { double_organisation(:update_attributes_with_admin => false) }
+          allow(Organisation).to receive(:find) { double_organisation(:update_attributes_with_superadmin => false) }
           put :update, :id => "1", :organisation => {'these' => 'params'}
           expect(response).to render_template("edit")
         end
@@ -467,9 +453,9 @@ describe OrganisationsController, :type => :controller do
   end
 
   describe "DELETE destroy" do
-    context "while signed in as admin", :helpers => :controllers do
+    context "while signed in as superadmin", :helpers => :controllers do
       before(:each) do
-        make_current_user_admin
+        make_current_user_superadmin
       end
       it "destroys the requested organisation and redirect to organisation list" do
         expect(Organisation).to receive(:find).with('37') { double_organisation }
@@ -478,9 +464,9 @@ describe OrganisationsController, :type => :controller do
         expect(response).to redirect_to(organisations_url)
       end
     end
-    context "while signed in as non-admin", :helpers => :controllers do
+    context "while signed in as non-superadmin", :helpers => :controllers do
       before(:each) do
-        make_current_user_nonadmin
+        make_current_user_nonsuperadmin
       end
       it "does not destroy the requested organisation but redirects to organisation home page" do
         double = double_organisation
