@@ -11,86 +11,63 @@ describe Organisation, :type => :model do
     @category3 = FactoryGirl.create(:category, :charity_commission_id => 108)
     @category4 = FactoryGirl.create(:category, :charity_commission_id => 302)
     @category5 = FactoryGirl.create(:category, :charity_commission_id => 306)
-    @org1 = FactoryGirl.build(:organisation, :email => nil, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 3TE', :donation_info => 'www.harrow-bereavment.co.uk/donate')
+    @org1 = FactoryGirl.build(:organisation, :email => "", :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '84 pinner road', :postcode => 'HA1 4HA', :donation_info => 'www.harrow-bereavment.co.uk/donate')
     @org1.save!
-    @org2 = FactoryGirl.build(:organisation, :email => nil,  :name => 'Indian Elders Association',
-                              :description => 'Care for the elderly', :address => '64 pinner road', :postcode => 'HA1 3RE', :donation_info => 'www.indian-elders.co.uk/donate')
+    @org2 = FactoryGirl.build(:organisation, :name => 'Indian Elders Association',:email => "",
+                              :description => 'Care for the elderly', :address => '64 pinner road', :postcode => 'HA1 4HZ', :donation_info => 'www.indian-elders.co.uk/donate')
     @org2.categories << @category1
     @org2.categories << @category2
+    @org2.categories << @category3
     @org2.save!
-    @org3 = FactoryGirl.build(:organisation, :email => nil, :name => 'Age UK Elderly', :description => 'Care for older people', :address => '64 pinner road', :postcode => 'HA1 3RE', :donation_info => 'www.age-uk.co.uk/donate')
+    @org3 = FactoryGirl.build(:organisation, :email => "", :name => 'Age UK Elderly', :description => 'Care for older people', :address => '64 pinner road', :postcode => 'HA1 4HZ', :donation_info => 'www.age-uk.co.uk/donate')
     @org3.categories << @category1
     @org3.save!
   end
 
-  describe '#not_updated_recently?' do
-    let(:org){FactoryGirl.create(:organisation, updated_at: Time.now)}
-
-    it{expect(org.not_updated_recently?).to be false}
-
-    context "updated too long ago" do
-      let(:org){FactoryGirl.create(:organisation, updated_at: 365.day.ago)}
-      it{expect(org.not_updated_recently?).to be true}
-    end
-
-    context "when updated recently" do
-      let(:org){FactoryGirl.create(:organisation, updated_at: 364.day.ago)}
-      it{expect(org.not_updated_recently?).to be false}
-    end
-  end
-
-  describe "#not_updated_recently_or_has_no_owner?" do
-    let(:subject){FactoryGirl.create(:organisation, :name => "Org with no owner", :updated_at => 364.day.ago)}
-    context 'has no owner but updated recently' do
-      it{expect(subject.not_updated_recently_or_has_no_owner?).to be true}
-    end
-    context 'has owner but old update' do
-      let(:subject){FactoryGirl.create(:organisation_with_owner, :updated_at => 366.day.ago)}
-      it{expect(subject.not_updated_recently_or_has_no_owner?).to be true}
-    end
-    context 'has no owner and old update' do 
-      let(:subject){FactoryGirl.create(:organisation, :updated_at => 366.day.ago)}
-      it{expect(subject.not_updated_recently_or_has_no_owner?).to be true}
-    end
-    context 'has owner and recent update' do
-      let(:subject){FactoryGirl.create(:organisation_with_owner, :updated_at => 364.day.ago)}
-      it{expect(subject.not_updated_recently_or_has_no_owner?).to be false}
-    end
-  end
-
   describe "#gmaps4rails_marker_attrs" do
+
+    def build_org_with_computed_fields_and_updated_at org, updated_at = nil
+      org.update_attributes(updated_at: updated_at) unless updated_at.nil?
+      Queries::Organisations.add_recently_updated_and_has_owner(Organisation.where(id: org.id)).first
+    end
     context 'no user' do
       it 'returns small icon when no associated user' do
-        expect(@org1.gmaps4rails_marker_attrs).to eq(["https://maps.gstatic.com/intl/en_ALL/mapfiles/markers2/measle.png", {"data-id"=>@org1.id, :class=>"measle"}])
+        expect(build_org_with_computed_fields_and_updated_at(@org1).gmaps4rails_marker_attrs).to eq(["https://maps.gstatic.com/intl/en_ALL/mapfiles/markers2/measle.png",
+          {"data-id"=>@org1.id, :class=>"measle"}])
       end
     end
 
     context 'has user' do
       before(:each) do
-        usr = FactoryGirl.create(:user, :email => "orgadmin@org.org")
+        usr = FactoryGirl.create(:user, :email => "orgsuperadmin@org.org")
         usr.confirm!
         @org1.users << [usr]
         @org1.save!
       end
-      after(:each) do
-        allow(Time).to receive(:now).and_call_original
-      end
       it 'returns large icon when there is an associated user' do
-        expect(@org1.gmaps4rails_marker_attrs).to eq( ["http://mt.googleapis.com/vt/icon/name=icons/spotlight/spotlight-poi.png", {"data-id"=>@org1.id,:class=>"marker"}])
+        expect(build_org_with_computed_fields_and_updated_at(@org1).gmaps4rails_marker_attrs).to eq( ["https://mt.googleapis.com/vt/icon/name=icons/spotlight/spotlight-poi.png",
+          {"data-id"=>@org1.id,:class=>"marker"}])
       end
 
-      [365, 366, 500].each do |days|
+      [ 365, 366, 500 ].each do |days|
         it "returns small icon when update is #{days} days old" do
-          future_time = Time.at(Time.now + days.day)
-          allow(Time).to receive(:now){future_time}
-          expect(@org1.gmaps4rails_marker_attrs).to eq(["https://maps.gstatic.com/intl/en_ALL/mapfiles/markers2/measle.png", {"data-id"=>@org1.id, :class=>"measle"}])
+          # adds generous 5 second pad for query to run
+          past_time = Time.current.advance(days: -days).advance(seconds: -5)
+          expect(
+            build_org_with_computed_fields_and_updated_at(
+              @org1, past_time
+            ).gmaps4rails_marker_attrs
+          ).to eq([
+            "https://maps.gstatic.com/intl/en_ALL/mapfiles/markers2/measle.png",
+            {"data-id"=>@org1.id, :class=>"measle"}
+          ])
         end
       end
       [ 2, 100, 200, 364].each do |days|
         it "returns large icon when update is only #{days} days old" do
-          future_time = Time.at(Time.now + days.day)
-          allow(Time).to receive(:now){future_time}
-          expect(@org1.gmaps4rails_marker_attrs).to eq( ["http://mt.googleapis.com/vt/icon/name=icons/spotlight/spotlight-poi.png", {"data-id"=>@org1.id, :class=>"marker"} ])
+          past_time = Time.at(Time.now - days.day)
+          expect(build_org_with_computed_fields_and_updated_at(@org1, past_time).gmaps4rails_marker_attrs).to eq( ["https://mt.googleapis.com/vt/icon/name=icons/spotlight/spotlight-poi.png",
+            {"data-id"=>@org1.id, :class=>"marker"} ])
         end
       end
     end
@@ -108,7 +85,7 @@ describe Organisation, :type => :model do
       expect(Organisation.not_null_email).to eq [@org1]
     end
 
-    it 'should allow us to grab orgs with no admin' do
+    it 'should allow us to grab orgs with no superadmin' do
       expect(Organisation.null_users.sort).to eq [@org1, @org2, @org3].sort
       @org1.email = "hello@hello.com"
       @org1.save
@@ -135,8 +112,8 @@ describe Organisation, :type => :model do
   end
 
   context 'validating URLs' do
-    subject(:no_http_org) { FactoryGirl.build(:organisation, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 3TE', :donation_info => 'www.harrow-bereavment.co.uk/donate') }
-    subject(:empty_website)  {FactoryGirl.build(:organisation, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 3TE', :donation_info => '', :website => '')}
+    subject(:no_http_org) { FactoryGirl.build(:organisation, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 4HZ', :donation_info => 'www.harrow-bereavment.co.uk/donate') }
+    subject(:empty_website)  {FactoryGirl.build(:organisation, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 4HZ', :donation_info => '', :website => '')}
     it 'if lacking protocol, http is prefixed to URL when saved' do
       no_http_org.save!
       expect(no_http_org.donation_info).to include('http://')
@@ -151,57 +128,72 @@ describe Organisation, :type => :model do
     end
   end
 
-  context 'adding charity admins by email' do
-    it 'handles a non-existent email with an error' do
-      expect(@org1.update_attributes_with_admin({:admin_email_to_add => 'nonexistentuser@example.com'})).to be_nil
-      expect(@org1.errors[:administrator_email]).to eq ["The user email you entered,'nonexistentuser@example.com', does not exist in the system"]
+  context 'adding charity superadmins by email' do
+    it 'handles an invalid email with an error' do
+      @org1.update_attributes_with_superadmin({:superadmin_email_to_add => 'user'})
+      expect(@org1.errors.messages[:superadministrator_email]).to include "The user email you entered,'user', is invalid"
     end
-    it 'does not update other attributes when there is a non-existent email' do
-      expect(@org1.update_attributes_with_admin({:name => 'New name',:admin_email_to_add => 'nonexistentuser@example.com'})).to be_nil
+    it 'does not email when email is invalid' do
+      expect{
+        @org1.update_attributes_with_superadmin({:superadmin_email_to_add => 'user'})
+      }.not_to change(ActionMailer::Base.deliveries, :length)
+    end
+    it 'does not update other attributes when email is invalid' do
+      @org1.update_attributes_with_superadmin({:superadmin_email_to_add => 'user', :name => "Random name"})
+      expect(@org1.name).not_to eq "Random name"
+    end
+
+    it 'handles a non-existent email by inviting user' do
+      expect(@org1.update_attributes_with_superadmin({:superadmin_email_to_add => 'nonexistentuser@example.com'})).to be true
+      expect(@org1).to be_valid
+      usr = User.find_by(email:'nonexistentuser@example.com')
+      expect(usr).not_to be_nil
+      expect(usr.organisation).to eq @org1
+    end
+    it 'does not update other attributes when there is an invalid email' do
+      expect(@org1.update_attributes_with_superadmin({:name => 'New name',:superadmin_email_to_add => 'user'})).to be_nil
       expect(@org1.name).not_to eq 'New name'
     end
     it 'handles a nil email' do
-      expect(@org1.update_attributes_with_admin({:admin_email_to_add => nil})).to be true
+      expect(@org1.update_attributes_with_superadmin({:superadmin_email_to_add => nil})).to be true
       expect(@org1.errors.any?).to be false
     end
     it 'handles a blank email' do
-      expect(@org1.update_attributes_with_admin({:admin_email_to_add => ''})).to be true
+      expect(@org1.update_attributes_with_superadmin({:superadmin_email_to_add => ''})).to be true
       expect(@org1.errors.any?).to be false
     end
-    it 'adds existent user as charity admin' do
+    it 'adds existent user as charity superadmin' do
       usr = FactoryGirl.create(:user, :email => 'user@example.org')
-      expect(@org1.update_attributes_with_admin({:admin_email_to_add => usr.email})).to be true
+      expect(@org1.update_attributes_with_superadmin({:superadmin_email_to_add => usr.email})).to be true
       expect(@org1.users).to include usr
     end
+    it 'uses org admin mailer to email existent user when upgraded to org admin' do
+      usr = FactoryGirl.create(:user, :email => 'user@example.org')
+      mockMessage = double("message")
+      expect(mockMessage).to receive :deliver_now
+      expect(OrgAdminMailer).to receive_message_chain(:new_org_admin).with(@org1, [usr.email]).and_return(mockMessage)
+      @org1.update_attributes_with_superadmin({:superadmin_email_to_add => usr.email})
+    end
     it 'updates other attributes with blank email' do
-      expect(@org1.update_attributes_with_admin({:name => 'New name',:admin_email_to_add => ''})).to be true
+      expect(@org1.update_attributes_with_superadmin({:name => 'New name',:superadmin_email_to_add => ''})).to be true
       expect(@org1.name).to eq 'New name'
     end
     it 'updates other attributes with valid email' do
       usr = FactoryGirl.create(:user, :email => 'user@example.org')
-      expect(@org1.update_attributes_with_admin({:name => 'New name',:admin_email_to_add => usr.email})).to be true
+      expect(@org1.update_attributes_with_superadmin({:name => 'New name',:superadmin_email_to_add => usr.email})).to be true
       expect(@org1.name).to eq 'New name'
     end
-  end
-  it 'responds to filter by category' do
-    expect(Organisation).to respond_to(:filter_by_category)
-  end
-
-  it 'finds all orgs in a particular category' do
-    expect(Organisation.filter_by_category(@category1.id)).not_to include @org1
-    expect(Organisation.filter_by_category(@category1.id)).to include @org2
-    expect(Organisation.filter_by_category(@category1.id)).to include @org3
-  end
-
-  it 'finds all orgs when category is nil' do
-    expect(Organisation.filter_by_category(nil)).to include(@org1)
-    expect(Organisation.filter_by_category(nil)).to include(@org2)
-    expect(Organisation.filter_by_category(nil)).to include(@org3)
   end
 
   it 'should have and belong to many categories' do
     expect(@org2.categories).to include(@category1)
     expect(@org2.categories).to include(@category2)
+  end
+
+  it 'should have and belong to many categories by types' do
+    expect(@org2.categories.what_they_do).to include(@category3)
+    expect(@org2.categories.who_they_help).to include(@category1)
+    expect(@org2.categories.how_they_help).to include(@category2)
   end
 
   it 'must have search by keyword' do
@@ -210,39 +202,6 @@ describe Organisation, :type => :model do
 
   it 'find all orgs that have keyword anywhere in their name or description' do
     expect(Organisation.search_by_keyword("elderly")).to eq([@org2, @org3])
-  end
-
-  it 'searches by keyword and filters by category and has zero results' do
-    result = Organisation.search_by_keyword("Harrow").filter_by_category("1")
-    expect(result).not_to include @org1, @org2, @org3
-  end
-
-  it 'searches by keyword and filters by category and has results' do
-    result = Organisation.search_by_keyword("Indian").filter_by_category(@category1.id)
-    expect(result).to include @org2
-    expect(result).not_to include @org1, @org3
-  end
-
-  it 'searches by keyword when filter by category id is nil' do
-    result = Organisation.search_by_keyword("Harrow").filter_by_category(nil)
-    expect(result).to include @org1
-    expect(result).not_to include @org2, @org3
-  end
-
-  it 'filters by category when searches by keyword is nil' do
-    result = Organisation.search_by_keyword(nil).filter_by_category(@category1.id)
-    expect(result).to include @org2, @org3
-    expect(result).not_to include @org1
-  end
-
-  it 'returns all orgs when both filter by category and search by keyword are nil args' do
-    result = Organisation.search_by_keyword(nil).filter_by_category(nil)
-    expect(result).to include @org1, @org2, @org3
-  end
-
-  it 'handles weird input (possibly from infinite scroll system)' do
-    # Couldn't find Category with id=?test=0
-    expect(lambda {Organisation.filter_by_category("?test=0")} ).not_to raise_error
   end
 
   it 'has users' do
@@ -254,10 +213,12 @@ describe Organisation, :type => :model do
   end
 
   describe 'Creating of Organisations from CSV file' do
-    before(:all){ @headers = 'Title,Charity Number,Activities,Contact Name,Contact Address,website,Contact Telephone,date registered,date removed,accounts date,spending,income,company number,OpenlyLocalURL,twitter account name,facebook account name,youtube account name,feed url,Charity Classification,signed up for 1010,last checked,created at,updated at,Removed?'.split(',')}
-
+    before(:each) do
+      allow_any_instance_of(Organisation).to receive(:geocode)
+      @headers = 'Title,Charity Number,Activities,Contact Name,Contact Address,website,Contact Telephone,date registered,date removed,accounts date,spending,income,company number,OpenlyLocalURL,twitter account name,facebook account name,youtube account name,feed url,Charity Classification,signed up for 1010,last checked,created at,updated at,Removed?'.split(',')
+    end
     it 'must not override an existing organisation' do
-      fields = CSV.parse('INDIAN ELDERS ASSOCIATION,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH,COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      fields = CSV.parse('INDIAN ELDERS ASSOCIATION,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH,COLLEGE ROAD, HARROW, HA1 4HZ",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
       org = create_organisation(fields)
       expect(org).to be_nil
     end
@@ -275,6 +236,7 @@ describe Organisation, :type => :model do
     end
 
     it 'must fail gracefully when encountering error in generating multiple Organisations from text file' do
+      allow(Organisation).to receive(:sleep)
       attempted_number_to_import = 1006
       actual_number_to_import = 642
       allow(Organisation).to receive(:create_from_array).and_raise(CSV::MalformedCSVError)
@@ -292,7 +254,7 @@ describe Organisation, :type => :model do
       expect(org.postcode).to eq('')
       expect(org.website).to eq('http://www.harrow-baptist.org.uk')
       expect(org.telephone).to eq('020 8863 7837')
-      expect(org.donation_info).to eq(nil)
+      expect(org.donation_info).to eq("")
     end
 
     it 'must be able to handle no address in text representation' do
@@ -304,26 +266,26 @@ describe Organisation, :type => :model do
       expect(org.postcode).to eq('')
       expect(org.website).to eq('http://www.harrow-baptist.org.uk')
       expect(org.telephone).to eq('020 8863 7837')
-      expect(org.donation_info).to eq(nil)
+      expect(org.donation_info).to eq("")
     end
 
     it 'must be able to generate Organisation from text representation ensuring words in correct case and postcode is extracted from address' do
-      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 4HZ",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
       org = create_organisation(fields)
       expect(org.name).to eq('Harrow Baptist Church')
       expect(org.description).to eq('No information recorded')
       expect(org.address).to eq('Harrow Baptist Church, College Road, Harrow')
-      expect(org.postcode).to eq('HA1 1BA')
+      expect(org.postcode).to eq('HA1 4HZ')
       expect(org.website).to eq('http://www.harrow-baptist.org.uk')
       expect(org.telephone).to eq('020 8863 7837')
-      expect(org.donation_info).to eq(nil)
+      expect(org.donation_info).to eq("")
     end
 
 
     it 'should raise error if no columns found' do
       #Headers are without Title header
       @headers = 'Charity Number,Activities,Contact Name,Contact Address,website,Contact Telephone,date registered,date removed,accounts date,spending,income,company number,OpenlyLocalURL,twitter account name,facebook account name,youtube account name,feed url,Charity Classification,signed up for 1010,last checked,created at,updated at,Removed?'.split(',')
-      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 4HZ",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
       expect(lambda{
         org = create_organisation(fields)
       }).to raise_error
@@ -337,19 +299,19 @@ describe Organisation, :type => :model do
 
     context "importing category relations" do
       let(:fields) do
-        CSV.parse('HARROW BEREAVEMENT COUNSELLING,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+        CSV.parse('HARROW BEREAVEMENT COUNSELLING,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 4HZ",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
       end
       let(:row) do
         CSV::Row.new(@headers, fields.flatten)
       end
       let(:fields_cat_missing) do
-        CSV.parse('HARROW BEREAVEMENT COUNSELLING,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,,false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+        CSV.parse('HARROW BEREAVEMENT COUNSELLING,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 4HZ",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,,false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
       end
       let(:row_cat_missing) do
         CSV::Row.new(@headers, fields_cat_missing.flatten)
       end
       it 'must be able to avoid org category relations from text file when org does not exist' do
-        @org4 = FactoryGirl.build(:organisation, :name => 'Fellowship For Management In Food Distribution', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 3TE', :donation_info => 'www.harrow-bereavment.co.uk/donate')
+        @org4 = FactoryGirl.build(:organisation, :name => 'Fellowship For Management In Food Distribution', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 4HZ', :donation_info => 'www.harrow-bereavment.co.uk/donate')
         @org4.save!
         [102,206,302].each do |id|
           FactoryGirl.build(:category, :charity_commission_id => id).save!
@@ -407,20 +369,6 @@ describe Organisation, :type => :model do
     end
   end
 
-  it 'should geocode when address changes' do
-    new_address = '60 pinner road'
-    expect(@org1).to receive(:geocode)
-    @org1.update_attributes :address => new_address
-  end
-
-  it 'should geocode when new object is created' do
-    address = '60 pinner road'
-    postcode = 'HA1 3RE'
-    org = FactoryGirl.build(:organisation,:address => address, :postcode => postcode, :name => 'Happy and Nice', :gmaps => true)
-    expect(org).to receive(:geocode)    
-    org.save
-  end
-  
   # not sure if we need SQL injection security tests like this ...
   # org = Organisation.new(:address =>"blah", :gmaps=> ";DROP DATABASE;")
   # org = Organisation.new(:address =>"blah", :name=> ";DROP DATABASE;")
@@ -452,6 +400,13 @@ describe Organisation, :type => :model do
       Organisation.add_email(fields = CSV.parse('friendly,,,,,,,test@example.org')[0],true)
     end
 
+    it "should add blank string email to org if email is NULL" do
+      expect(Organisation).to receive(:where).with("UPPER(name) LIKE ? ", "%FRIENDLY%").and_return([@org1])
+      expect{
+        Organisation.add_email(fields = CSV.parse('friendly,,,,,,,')[0],true)
+      }.not_to change(@org1, :email)
+    end
+
     it "should add email to org even with case mismatch" do
       expect(Organisation).to receive(:where).with("UPPER(name) LIKE ? ", "%FRIENDLY%").and_return([@org1])
       expect(@org1).to receive(:email=).with('test@example.org')
@@ -479,7 +434,7 @@ describe Organisation, :type => :model do
   end
 
   describe '#uninvite_users' do
-    let!(:current_user) { FactoryGirl.create(:user, email: 'admin@example.com', admin: true) }
+    let!(:current_user) { FactoryGirl.create(:user, email: 'superadmin@example.com', superadmin: true) }
     let(:org) { FactoryGirl.create :organisation, email: 'YES@hello.com' }
     let(:params) do
       {invite_list: {org.id => org.email,
@@ -586,11 +541,11 @@ describe Organisation, :type => :model do
     FactoryGirl.factories.clear
     FactoryGirl.find_definitions
 
-    @org1 = FactoryGirl.build(:organisation, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 3TE', :donation_info => 'www.harrow-bereavment.co.uk/donate')
+    @org1 = FactoryGirl.build(:organisation, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 4HZ', :donation_info => 'www.harrow-bereavment.co.uk/donate')
     @org1.save!
-    @org2 = FactoryGirl.build(:organisation, :name => 'Indian Elders Associaton', :description => 'Care for the elderly', :address => '64 pinner road', :postcode => 'HA1 3RE', :latitude => 77, :longitude => 77, :donation_info => 'www.indian-elders.co.uk/donate')
+    @org2 = FactoryGirl.build(:organisation, :name => 'Indian Elders Associaton', :description => 'Care for the elderly', :address => '64 pinner road', :postcode => 'HA1 4HZ', :latitude => 77, :longitude => 77, :donation_info => 'www.indian-elders.co.uk/donate')
     @org2.save!
-    @org3 = FactoryGirl.build(:organisation, :name => 'Age UK Elderly', :description => 'Care for older people', :address => '64 pinner road', :postcode => 'HA1 3RE', :donation_info => 'www.age-uk.co.uk/donate')
+    @org3 = FactoryGirl.build(:organisation, :name => 'Age UK Elderly', :description => 'Care for older people', :address => '64 pinner road', :postcode => 'HA1 4HZ', :donation_info => 'www.age-uk.co.uk/donate')
     @org3.save!
   end
   
@@ -654,4 +609,115 @@ describe Organisation, :type => :model do
       @org2.save!
     end
   end
+end
+
+describe Organisation, '::filter_by_categories' do
+  let!(:category1) { create(:category, :charity_commission_id => 108) }
+  let!(:category2) { create(:category, :charity_commission_id => 205) }
+  # not initialized!
+  let(:category3) { create(:category, :charity_commission_id => 307) }
+
+  let!(:org1) do
+    create(
+      :organisation,
+      :email => "",
+      :name => 'Harrow Bereavement Counselling',
+      :description => 'Bereavement Counselling',
+      :address => '64 pinner road',
+      :postcode => 'HA1 4HZ',
+      :donation_info => 'www.harrow-bereavment.co.uk/donate',
+    )
+  end
+
+  let!(:org2) do
+    create(
+      :organisation,
+      :name => 'Indian Elders Association',
+      :email => "",
+      :description => 'Care for the elderly',
+      :address => '64 pinner road',
+      :postcode => 'HA1 4HZ',
+      :donation_info => 'www.indian-elders.co.uk/donate'
+    ).tap { |o| o.categories << category1 ; o.categories << category2 }
+  end
+
+  let!(:org3) do
+    create(
+      :organisation,
+      :email => "",
+      :name => 'Age UK Elderly',
+      :description => 'Care for older people',
+      :address => '64 pinner road',
+      :postcode => 'HA1 4HZ',
+      :donation_info => 'www.age-uk.co.uk/donate',
+    ).tap { |o| o.categories  << category1 }
+  end
+
+  context 'when filtering by ONE CATEGORY, it returns only organisations that
+           \ are associated with that category' do
+
+    it 'organisations returned by query' do
+      expect(
+        Organisation.filter_by_categories([category1.id]).pluck(:id)
+      ).to include(
+        org2.id, org3.id
+      )
+    end
+
+    it 'no duplicates' do
+      expect(
+        Organisation.filter_by_categories([category1.id]).map.size
+      ).to eq 2
+    end
+
+    it 'categories in join table' do
+      expect(
+        CategoryOrganisation.where(
+          organisation_id: Organisation.filter_by_categories([category1.id]).select(:id)
+        ).pluck(:category_id).uniq
+      ).to include(
+        category1.id, category2.id
+      )
+    end
+  end
+
+  context 'when filtering by TWO CATEGORIES, it returns only organisations that
+           \ are associated with those categories' do
+    # init
+    before { category3 }
+
+    it 'organisations returned by query' do
+      expect(
+        Organisation.filter_by_categories([
+          category1.id,
+          category2.id,
+        ]).pluck(:id)
+      ).to include(
+        org2.id
+      )
+    end
+
+    it 'no duplicates' do
+      expect(
+        Organisation.filter_by_categories([
+          category1.id,
+          category2.id,
+        ]).map.size
+      ).to eq 1
+    end
+
+    it 'categories in join table' do
+      expect(
+        CategoryOrganisation.where(
+          organisation_id: Organisation.filter_by_categories([
+            category1.id,
+            category2.id,
+          ]).select(:id)
+        ).pluck(:category_id).uniq
+      ).to include(
+        category1.id, category2.id
+      )
+    end
+  end
+
 end
