@@ -1,40 +1,41 @@
-class OrganisationsController < ApplicationController
-  layout 'two_columns'
+class OrganisationsController < BaseOrganisationsController
+  layout 'two_columns_with_map'
   # GET /organisations/search
   # GET /organisations/search.json
   before_filter :authenticate_user!, :except => [:search, :index, :show]
 
   def search
-    @query_term = params[:q]
-    @category_id = params.try(:[],'category').try(:[],'id')
-    @category = Category.find_by_id(@category_id)
-    @organisations = Organisation.includes(:users).order_by_most_recent
-    @organisations = @organisations.search_by_keyword(@query_term).filter_by_category(@category_id)
+    @parsed_params = SearchParamsParser.new(params)
+    @cat_name_ids = Category.name_and_id_for_what_who_and_how
+    @organisations = Queries::Organisations.search_by_keyword_and_category(
+      @parsed_params
+    )
     flash.now[:alert] = SEARCH_NOT_FOUND if @organisations.empty?
     @markers = build_map_markers(@organisations)
-    @category_options = Category.html_drop_down_options
+
     render :template =>'organisations/index'
   end
 
   # GET /organisations
   # GET /organisations.json
   def index
-    @organisations = Organisation.includes(:users).order_by_most_recent
+    @organisations = Queries::Organisations.order_by_most_recent
     @markers = build_map_markers(@organisations)
-    @category_options = Category.html_drop_down_options
+    @cat_name_ids = Category.name_and_id_for_what_who_and_how
   end
 
   # GET /organisations/1
   # GET /organisations/1.json
   def show
-    @organisation = Organisation.find(params[:id])
-    @pending_admin = current_user.pending_admin? @organisation if current_user
+    organisations = Organisation.where(id: params[:id])
+    @organisation = organisations.first!
+    @pending_org_admin = current_user.pending_org_admin? @organisation if current_user
     @editable = current_user.can_edit?(@organisation) if current_user
     @deletable = current_user.can_delete?(@organisation) if current_user
     @can_create_volunteer_op = current_user.can_create_volunteer_ops?(@organisation) if current_user
     @grabbable = current_user ? current_user.can_request_org_admin?(@organisation) : true
-   # @next_path = current_user ? organisation_user_path(@organisation.id, current_user.id) : new_user_session_path
-    @markers = build_map_markers([@organisation])
+    @can_propose_edits = current_user.present? && !@editable
+    @markers = build_map_markers(organisations)
   end
 
   # GET /organisations/new
@@ -46,8 +47,9 @@ class OrganisationsController < ApplicationController
 
   # GET /organisations/1/edit
   def edit
-    @organisation = Organisation.find(params[:id])
-    @markers = build_map_markers([@organisation])
+    organisations = Organisation.where(id: params[:id])
+    @organisation = organisations.first!
+    @markers = build_map_markers(organisations)
     @categories_start_with = Category.first_category_name_in_each_type
     return false unless user_can_edit? @organisation
     #respond_to do |format|
@@ -58,10 +60,10 @@ class OrganisationsController < ApplicationController
   # POST /organisations
   # POST /organisations.json
   def create
-    # model filters for logged in users, but we check here if that user is an admin
+    # model filters for logged in users, but we check here if that user is an superadmin
     # TODO refactor that to model responsibility?
      org_params = OrganisationParams.build params
-     unless current_user.try(:admin?)
+     unless current_user.try(:superadmin?)
        flash[:notice] = PERMISSION_DENIED
        redirect_to organisations_path and return false
      end
@@ -78,13 +80,14 @@ class OrganisationsController < ApplicationController
   # PUT /organisations/1.json
   def update
     @organisation = Organisation.find(params[:id])
-    params[:organisation][:admin_email_to_add] = params[:organisation_admin_email_to_add] if params[:organisation]
+    params[:organisation][:superadmin_email_to_add] = params[:organisation_superadmin_email_to_add] if params[:organisation]
     update_params = OrganisationParams.build params 
     return false unless user_can_edit? @organisation
-    if @organisation.update_attributes_with_admin(update_params)
+    if @organisation.update_attributes_with_superadmin(update_params)
       redirect_to @organisation, notice: 'Organisation was successfully updated.'
     else
-      flash[:error] = @organisation.errors[:administrator_email][0]
+      @categories_start_with = Category.first_category_name_in_each_type
+      flash[:error] = @organisation.errors[:superadministrator_email][0]
       render action: "edit"
     end
   end
@@ -92,7 +95,7 @@ class OrganisationsController < ApplicationController
   # DELETE /organisations/1
   # DELETE /organisations/1.json
   def destroy
-    unless current_user.try(:admin?)
+    unless current_user.try(:superadmin?)
       flash[:notice] = PERMISSION_DENIED
       redirect_to organisation_path(params[:id]) and return false
     end
@@ -103,31 +106,27 @@ class OrganisationsController < ApplicationController
     redirect_to organisations_path
   end
 
-class OrganisationParams 
+  class OrganisationParams
     def self.build params
-      params.require(:organisation).permit( :admin_email_to_add, :description, :address, :publish_address, :postcode, :email, 
-                     :publish_email, :website, :publish_phone, :donation_info, :name, :telephone,
-                     category_organisations_attributes: [:_destroy, :category_id, :id])
+      params.require(:organisation).permit(
+        :superadmin_email_to_add,
+        :description,
+        :address,
+        :publish_address,
+        :postcode,
+        :email,
+        :publish_email,
+        :website,
+        :publish_phone,
+        :donation_info,
+        :name,
+        :telephone,
+        category_organisations_attributes: [:_destroy, :category_id, :id]
+      )
     end
   end
 
   private
-
-  def build_map_markers(organisations)
-    ::MapMarkerJson.build(organisations) do |org, marker|
-      marker.lat org.latitude
-      marker.lng org.longitude
-      marker.infowindow render_to_string(partial: 'popup', locals: {org: org})
-      marker.json(
-        custom_marker: render_to_string(
-          partial: 'shared/custom_marker',
-          locals: { attrs: org.gmaps4rails_marker_attrs }
-        ),
-        index: org.not_updated_recently_or_has_no_owner? ? -1 : 1,
-        type:  org.not_updated_recently_or_has_no_owner? ? 'small_org' : 'large_org'
-      )
-    end
-  end
 
   def user_can_edit?(org)
     unless current_user.try(:can_edit?,org)
@@ -136,4 +135,5 @@ class OrganisationParams
     end
     true
   end
+
 end
