@@ -5,7 +5,7 @@ class UsersController < ApplicationController
   end
 
   # we should allow logged in users to update their pending_organisation_id
-  # and only admins should be allowed to update organisation_id
+  # and only superadmins should be allowed to update organisation_id
   # makes me think of a attributes permissions matrix
   def update
     usr = User.find_by_id UserParams.build(params).fetch(:id)
@@ -15,13 +15,13 @@ class UsersController < ApplicationController
   def update_message_for_admin_status
     org = Organisation.find(UserParams.build(params).fetch(:pending_organisation_id))
     flash[:notice] = "You have requested admin status for #{org.name}"
-    send_email_to_site_admin_about_request_for_admin_of org   # could be moved to an hook on the user model?
+    send_email_to_superadmin_about_request_for_admin_of org   # could be moved to an hook on the user model?
     redirect_to organisation_path(UserParams.build(params).fetch(:pending_organisation_id))
   end
 
-  def send_email_to_site_admin_about_request_for_admin_of org
-    admin_emails = User.admins.pluck(:email)
-    AdminMailer.new_user_waiting_for_approval(org.name, admin_emails).deliver
+  def send_email_to_superadmin_about_request_for_admin_of org
+    superadmin_emails = User.superadmins.pluck(:email)
+    AdminMailer.new_user_waiting_for_approval(org.name, superadmin_emails).deliver_now
   end
   class UserParams
     def self.build params
