@@ -19,16 +19,16 @@ class User < ActiveRecord::Base
   belongs_to :pending_organisation, :class_name => 'Organisation', :foreign_key => 'pending_organisation_id'
 
   # should we have a before_save here where we check if the pending_organization_id is going from
-  # nil to a value and then send the admin an email ...
+  # nil to a value and then send the superadmin an email ...
 
   scope :invited_not_accepted,-> {includes(:organisation).where('users.invitation_sent_at IS NOT NULL').where('users.invitation_accepted_at IS NULL')}
-  scope :admins, -> { where(admin: true) }
+  scope :superadmins, -> { where(superadmin: true) }
 
   def can_create_volunteer_ops? org
-    belongs_to?(org) || admin?
+    belongs_to?(org) || superadmin?
   end
 
-  def pending_admin? org
+  def pending_org_admin? org
     return false if self.pending_organisation == nil
     self.pending_organisation == org
   end
@@ -44,16 +44,16 @@ class User < ActiveRecord::Base
 
   # can create or edit an organization
   def can_edit? org
-    admin? || (!org.nil? && organisation == org)
+    superadmin? || (!org.nil? && organisation == org)
   end
 
   def can_delete? org
-    admin?
+    superadmin?
   end
 
   def can_request_org_admin? org
-    # admin false, pending_organisation  pending_organisation!=organisation org != organisation
-    !admin? && organisation != org && pending_organisation != org
+    # superadmin false, pending_organisation  pending_organisation!=organisation org != organisation
+    !superadmin? && organisation != org && pending_organisation != org
   end
 
   def make_admin_of_org_with_matching_email
