@@ -8,15 +8,15 @@ class UserOrganisationClaimer
 
   def call(organisation_id)
     request_admin_status_if(organisation_id)
-    error_message_if_not_admin_or_not(organisation_id)
-    promote_user_if_admin_and_not(organisation_id)
+    error_message_if_not_superadmin_or_not(organisation_id)
+    promote_user_if_superadmin_and_not(organisation_id)
   end
 
   private
   attr_reader :listener, :user, :current_user
 
-  def is_current_user_admin?
-    current_user.admin?
+  def is_current_user_superadmin?
+    current_user.superadmin?
   end
 
   def request_admin_status_if(organisation_id)
@@ -26,14 +26,14 @@ class UserOrganisationClaimer
     end
   end
 
-  def error_message_if_not_admin_or_not(organisation_id)
-    if !is_current_user_admin? && !organisation_id
+  def error_message_if_not_superadmin_or_not(organisation_id)
+    if !is_current_user_superadmin? && !organisation_id
       listener.update_failure
     end
   end
 
-  def promote_user_if_admin_and_not(organisation_id)
-    if is_current_user_admin? && !organisation_id
+  def promote_user_if_superadmin_and_not(organisation_id)
+    if is_current_user_superadmin? && !organisation_id
       user.promote_to_org_admin
       listener.update_message_promoting(user)
     end
