@@ -8,7 +8,7 @@ LocalSupport::Application.configure do
   config.cache_classes = true
 
   # Configure static asset server for tests with Cache-Control for performance
-  config.serve_static_assets = true
+  config.serve_static_files = true
   config.static_cache_control = "public, max-age=3600"
 
   # Show full error reports and disable caching
@@ -37,7 +37,8 @@ LocalSupport::Application.configure do
   # Allow pass debug_assets=true as a query parameter to load pages with unpackaged assets
   config.assets.allow_debugging = true
 
-  config.action_mailer.default_url_options = { :host => 'localhost:3000' } 
+  config.action_mailer.default_url_options = { :host => 'localhost:3000', :port => 3000 }
+  Rails.application.routes.default_url_options[:host] = 'localhost:3000'
   config.middleware.use RackSessionAccess::Middleware
   config.eager_load = false
 end
