@@ -11,12 +11,12 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20141024215924) do
+ActiveRecord::Schema.define(version: 20150627195630) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "categories", force: true do |t|
+  create_table "categories", force: :cascade do |t|
     t.string   "name"
     t.integer  "charity_commission_id"
     t.string   "charity_commission_name"
@@ -24,37 +24,42 @@ ActiveRecord::Schema.define(version: 20141024215924) do
     t.datetime "updated_at"
   end
 
-  create_table "categories_organisations", force: true do |t|
+  create_table "categories_organisations", force: :cascade do |t|
     t.integer "category_id"
     t.integer "organisation_id"
   end
 
-  create_table "features", force: true do |t|
+  add_index "categories_organisations", ["category_id"], name: "index_categories_organisations_on_category_id", using: :btree
+  add_index "categories_organisations", ["organisation_id"], name: "index_categories_organisations_on_organisation_id", using: :btree
+
+  create_table "features", force: :cascade do |t|
     t.string  "name"
     t.boolean "active", default: false
   end
 
-  create_table "organisations", force: true do |t|
-    t.string   "name"
-    t.string   "address"
-    t.string   "postcode"
-    t.string   "email"
-    t.text     "description"
-    t.string   "website"
-    t.string   "telephone"
+  create_table "organisations", force: :cascade do |t|
+    t.string   "name",            default: "",             null: false
+    t.string   "address",         default: "",             null: false
+    t.string   "postcode",        default: "",             null: false
+    t.string   "email",           default: "",             null: false
+    t.text     "description",     default: "",             null: false
+    t.string   "website",         default: "",             null: false
+    t.string   "telephone",       default: "",             null: false
     t.datetime "created_at"
     t.datetime "updated_at"
     t.float    "latitude"
     t.float    "longitude"
     t.boolean  "gmaps"
-    t.text     "donation_info"
+    t.text     "donation_info",   default: "",             null: false
     t.boolean  "publish_address", default: false
     t.boolean  "publish_phone",   default: false
     t.boolean  "publish_email",   default: true
     t.datetime "deleted_at"
+    t.string   "type",            default: "Organisation"
+    t.boolean  "non_profit"
   end
 
-  create_table "pages", force: true do |t|
+  create_table "pages", force: :cascade do |t|
     t.string   "name"
     t.string   "permalink"
     t.text     "content"
@@ -65,7 +70,28 @@ ActiveRecord::Schema.define(version: 20141024215924) do
 
   add_index "pages", ["permalink"], name: "index_pages_on_permalink", using: :btree
 
-  create_table "users", force: true do |t|
+  create_table "proposed_organisation_edits", force: :cascade do |t|
+    t.integer  "organisation_id"
+    t.string   "name",            default: "",    null: false
+    t.string   "address",         default: "",    null: false
+    t.string   "postcode",        default: "",    null: false
+    t.string   "email",           default: "",    null: false
+    t.text     "description",     default: "",    null: false
+    t.string   "website",         default: "",    null: false
+    t.string   "telephone",       default: "",    null: false
+    t.text     "donation_info",   default: "",    null: false
+    t.datetime "deleted_at"
+    t.integer  "user_id"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.boolean  "accepted",        default: false, null: false
+    t.boolean  "archived",        default: false, null: false
+  end
+
+  add_index "proposed_organisation_edits", ["deleted_at"], name: "index_proposed_organisation_edits_on_deleted_at", using: :btree
+  add_index "proposed_organisation_edits", ["user_id"], name: "index_proposed_organisation_edits_on_user_id", using: :btree
+
+  create_table "users", force: :cascade do |t|
     t.string   "email",                   default: "",    null: false
     t.string   "encrypted_password",      default: ""
     t.string   "reset_password_token"
@@ -78,7 +104,7 @@ ActiveRecord::Schema.define(version: 20141024215924) do
     t.string   "last_sign_in_ip"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.boolean  "admin",                   default: false
+    t.boolean  "superadmin",              default: false
     t.integer  "organisation_id"
     t.string   "confirmation_token"
     t.datetime "confirmed_at"
@@ -93,14 +119,16 @@ ActiveRecord::Schema.define(version: 20141024215924) do
     t.integer  "invited_by_id"
     t.string   "invited_by_type"
     t.datetime "deleted_at"
+    t.boolean  "siteadmin",               default: false
   end
 
   add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
   add_index "users", ["invitation_token"], name: "index_users_on_invitation_token", unique: true, using: :btree
   add_index "users", ["invited_by_id"], name: "index_users_on_invited_by_id", using: :btree
+  add_index "users", ["organisation_id"], name: "index_users_on_organisation_id", using: :btree
   add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
 
-  create_table "volunteer_ops", force: true do |t|
+  create_table "volunteer_ops", force: :cascade do |t|
     t.string   "title"
     t.text     "description"
     t.integer  "organisation_id"
