@@ -7,10 +7,24 @@
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
 Organisation.import_addresses 'db/data.csv', 1006
+user = User.where(email: "superadmin@harrowcn.org.uk").first_or_initialize
+user.password = "asdf1234"
+user.password_confirmation = "asdf1234"
+user.confirmed_at = DateTime.now
+user.superadmin = true
+user.save!
+
 user = User.where(email: "admin@harrowcn.org.uk").first_or_initialize
 user.password = "asdf1234"
 user.password_confirmation = "asdf1234"
 user.confirmed_at = DateTime.now
-user.admin = true
+user.superadmin = true
+user.save!
+
+user = User.where(email: "siteadmin@harrowcn.org.uk").first_or_initialize
+user.password = "asdf1234"
+user.password_confirmation = "asdf1234"
+user.confirmed_at = DateTime.now
+user.siteadmin = true
 user.save!
 
