@@ -16,8 +16,8 @@ Then(/^all the checkboxes should be (un)?checked$/) do |negate|
   all('input[type="checkbox"]').each { |box| box.send(expectation, be_checked) }
 end
 
-Given(/^the admin invited a user for "(.*?)"$/) do |organisation_name|
-  current_user = User.find_by_admin true
+Given(/^the superadmin invited a user for "(.*?)"$/) do |organisation_name|
+  current_user = User.find_by_superadmin true
   org = Organisation.find_by_name(organisation_name)
   params = {
       :resend_invitation => 'true',
