@@ -1,3 +1,6 @@
+require 'coveralls'
+Coveralls.wear_merged!('rails')
+
 ENV['CUCUMBER'] = 'cucumber'
 
 # IMPORTANT: This file is generated by cucumber-rails - edit at your own peril.
@@ -13,7 +16,8 @@ require 'cucumber/rspec/doubles'
 require 'rack_session_access/capybara'
 require 'factory_girl_rails'
 require 'aruba/cucumber'
-require 'delorean'
+require 'timecop'
+require 'billy/cucumber'
 Dir['../../spec/factories/*.rb'].each {|file| require_relative file }
 
 # https://github.com/jnicklas/capybara/commit/4f805d5a1c42da29ed32ab0371e24add2dc08af1
@@ -25,9 +29,13 @@ end
 # prefer to use XPath just remove this line and adjust any selectors in your
 # steps to use the XPath syntax.
 Capybara.default_selector = :css
-Capybara.default_wait_time = 20
+Capybara.default_wait_time = 3
 
 Capybara.javascript_driver = :webkit
+Capybara::Webkit.configure do |config|
+	#config.debug = true
+	config.block_unknown_urls
+end
 
 # By default, any exception happening in your Rails application will bubble up
 # to Cucumber so that your scenario will fail. This is a different from how 
@@ -94,5 +102,10 @@ After('@in-production') do
 end
 
 After('@time_travel') do
-  Delorean.back_to_1985
+  Timecop.return
 end
+
+After do 
+	Capybara.use_default_driver
+end
+
