@@ -35,4 +35,50 @@ module Helpers
   end
 end
 
-World(Helpers)
+module MapHelpers
+  def stub_request_with_address(address, body = nil)
+    filename = "#{address.downcase.gsub(/,/,'').gsub(/\s/, '_')}.json"
+    filename = File.read "test/fixtures/#{filename}"
+    stub_request(:any, /maps\.googleapis\.com/).
+        to_return(status => 200, :body => body || filename, :headers => {})
+  end
+end
+
+module ProposedOrgHelpers
+  def unsaved_proposed_organisation(associated_user = nil)
+    proposed_org = ProposedOrganisation.new({name: "Friendly Charity", description: "We are friendly!",
+      email: "sample@sample.org", address: "30 pinner road", donation_info: "https://www.donate.com",
+      postcode: 'HA1 4HZ', non_profit: true})
+    proposed_org.users << associated_user if associated_user
+    proposed_org
+  end
+  def proposed_org_categories
+    [ 'Animal welfare',
+      'Accommodation',
+      'Education',
+      'Give them things' ]
+  end
+
+  def proposed_org_fields
+    {
+      name: 'Friendly charity',
+      address: '64 pinner road',
+      description: 'Such friendly so charity',
+      postcode: 'HA1 4HZ'
+    }
+  end
+
+  def visit_proposed_organisation opts
+    visit proposed_organisation_path(ProposedOrganisation.find_by(name: opts.fetch(:name)))
+  end
+
+  def press_acceptance_for_proposed_organisation
+    click_button 'Accept'
+  end
+
+  def assert_on_organisation_show_page opts
+    expect(current_path).to eq organisation_path(Organisation.find_by(name: opts.fetch(:name)))
+  end
+end
+
+World(Helpers, MapHelpers, ProposedOrgHelpers)
