@@ -1,18 +1,13 @@
 require 'rails_helper'
 
 describe SessionsController, :type => :controller do
-  before :suite do
-    FactoryGirl.factories.clear
-    FactoryGirl.find_definitions
-  end
-
   describe "POST create" do
     before :each do
       request.env["devise.mapping"] = Devise.mappings[:user]
     end
 
-    it 'redirects to home page after admin logs-in' do
-      FactoryGirl.build(:user, {:email => 'example@example.com', :password => 'pppppppp', :admin => true}).save!
+    it 'redirects to home page after superadmin logs-in' do
+      FactoryGirl.build(:user, {:email => 'example@example.com', :password => 'pppppppp', :superadmin => true}).save!
       post :create, 'user' => {'email' => 'example@example.com', 'password' => 'pppppppp'}
       expect(response).to redirect_to root_url
     end
@@ -27,7 +22,7 @@ describe SessionsController, :type => :controller do
       expect(flash[:alert]).to have_content "I'm sorry, you are not authorized to login to the system."
     end
 
-    it 'redirects to charity page after non-admin associated with org' do
+    it 'redirects to charity page after non-superadmin associated with org' do
       usr = FactoryGirl.build(:user_stubbed_organisation, {:email => 'example@example.com', :password => 'pppppppp'})
       allow(controller).to receive_messages :session => {previous_url: "/"}
       post :create, 'user' => {'email' => 'example@example.com', 'password' => 'pppppppp'}
