@@ -2,7 +2,7 @@ require 'rails_helper'
 describe UserReportsController, :type => :controller do
   describe 'PUT update user-organisation status', :helpers => :controllers do
     before(:each) do
-      make_current_user_admin
+      make_current_user_superadmin
       @nonadmin_user = double("User")
       allow(User).to receive(:find_by_id).with("4").and_return(@nonadmin_user)
       allow(@nonadmin_user).to receive(:pending_organisation_id=).with('5')
@@ -27,13 +27,13 @@ describe UserReportsController, :type => :controller do
         expect(flash[:notice]).to have_content("You have requested admin status for #{@org.name}")
       end
     end
-    context 'admin promoting user to charity admin' do
+    context 'superadmin promoting user to charity admin' do
       before(:each) do
         allow(@nonadmin_user).to receive(:promote_to_org_admin)
         allow(@nonadmin_user).to receive(:email).and_return('stuff@stuff.com')
       end
-      it 'non-admins get refused' do
-        allow(@nonadmin_user).to receive(:admin?).and_return(false)
+      it 'non-superadmins get refused' do
+        allow(@nonadmin_user).to receive(:superadmin?).and_return(false)
         allow(controller).to receive(:current_user).and_return(@nonadmin_user)
         put :update, {:id => '4'}
         expect(response.response_code).to eq(404)
@@ -54,14 +54,14 @@ describe UserReportsController, :type => :controller do
     let!(:user) { create :user }
 
     it 'destroys the user' do
-      make_current_user_admin
+      make_current_user_superadmin
       expect{
         delete :destroy, id: user.id
       }.to change(User, :count).by -1
     end
 
     it 'unless that user is the current_user' do
-      make_current_user_admin(user)
+      make_current_user_superadmin(user)
       expect{
         delete :destroy, id: user.id
       }.not_to change(User, :count)
@@ -70,9 +70,9 @@ describe UserReportsController, :type => :controller do
 
   describe 'GET index to view pending users' do
     context "user signed in", :helpers => :controllers do
-      context "as admin" do
+      context "as superadmin" do
         before(:each) do
-          make_current_user_admin
+          make_current_user_superadmin
         end
 
         it "assigns all users to @users" do
@@ -93,9 +93,9 @@ describe UserReportsController, :type => :controller do
         end
       end
 
-      context "as non-admin" do
+      context "as non-superadmin" do
         before(:each) do
-          make_current_user_nonadmin
+          make_current_user_nonsuperadmin
         end
 
         it "redirects user to root and flashes a notice" do
@@ -105,7 +105,7 @@ describe UserReportsController, :type => :controller do
 
         it "flashes the relevant notice" do
           get :index
-          expect(flash[:error]).to have_content("You must be signed in as an admin to perform this action!")
+          expect(flash[:error]).to have_content("You must be signed in as a superadmin to perform this action!")
         end
       end
     end
@@ -122,7 +122,7 @@ describe UserReportsController, :type => :controller do
 
       it "flashes the relevant notice" do
         get :index
-        expect(flash[:error]).to have_content("You must be signed in as an admin to perform this action!")
+        expect(flash[:error]).to have_content("You must be signed in as a superadmin to perform this action!")
       end
 
     end
@@ -145,12 +145,12 @@ describe UserReportsController, :type => :controller do
     end
 
     before do
-      make_current_user_admin
+      make_current_user_superadmin
       allow(User).to receive(:invited_not_accepted) { [user] }
     end
 
-    it 'is for admins only' do
-      make_current_user_nonadmin
+    it 'is for superadmins only' do
+      make_current_user_nonsuperadmin
       get :invited
       expect(response).to redirect_to root_path
     end
