@@ -4,7 +4,7 @@ describe UsersController, :type => :controller do
   describe ".permit" do
     it "returns the cleaned params" do
       user_params = {id: 5, email: "a@a.com", password: "blahblah", password_confirmation: "blahblah", 
-       remember_me: true, pending_organisation_id: 5, admin: true}
+       remember_me: true, pending_organisation_id: 5, superadmin: true}
       params = ActionController::Parameters.new.merge(user_params)
       permitted_params = UsersController::UserParams.build(params)
       expect(permitted_params).to eq({id: 5, email: "a@a.com", password: "blahblah", password_confirmation: "blahblah", 
