@@ -1,37 +1,34 @@
 describe('LocalSupport.InfoWindowBuilderFactory', function() {
-  var factory, google;
+    var factory, google;
 
-  beforeEach(function() {
-    google = {
-      maps: {
-        Size: function(x, y) {
-          this.msg = 'I am just a mock object!'
-          this.x = x;
-          this.y = y;
+    beforeEach(function() {
+      google = {
+        maps: {
+          Size: function(x, y) {
+            this.msg = 'I am just a mock object!'
+            this.x = x;
+            this.y = y;
+          }
         }
-      }
-    };
-    factory = LocalSupport.InfoWindowBuilderFactory;
-  });
+      };
+      factory = LocalSupport.InfoWindowBuilderFactory;
+    });
 
-  it('Volops infobox', function() {
-    expect(factory(google, 'vol_op').klass()).toEqual('arrow_box_vol_op')
-    expect(factory(google, 'vol_op').infobox('hi').content).toEqual('hi')
-    expect(factory(google, 'vol_op').infobox('hi').pixelOffset.x).toEqual(-151)
-    expect(factory(google, 'vol_op').infobox('hi').pixelOffset.y).toEqual(10)
-  });
+    it('Volops infobox', function() {
+      expect(factory(google, 'vol_op').infobox('hi').content).toEqual('hi')
+      expect(factory(google, 'vol_op').infobox('hi').pixelOffset.x).toEqual(-151)
+      expect(factory(google, 'vol_op').infobox('hi').pixelOffset.y).toEqual(10)
+    });
 
-  it('Small org infobox', function() {
-    expect(factory(google, 'small_org').klass()).toEqual('arrow_box')
-    expect(factory(google, 'small_org').infobox('hi').content).toEqual('hi')
-    expect(factory(google, 'small_org').infobox('hi').pixelOffset.x).toEqual(-151)
-    expect(factory(google, 'small_org').infobox('hi').pixelOffset.y).toEqual(-84)
-  });
+    it('Small org infobox', function() {
+      expect(factory(google, 'small_org').infobox('hi').content).toEqual('hi')
+      expect(factory(google, 'small_org').infobox('hi').pixelOffset.x).toEqual(-151)
+      expect(factory(google, 'small_org').infobox('hi').pixelOffset.y).toEqual(5)
+    });
 
-  it('Large org infobox', function() {
-    expect(factory(google, 'large_org').klass()).toEqual('arrow_box')
-    expect(factory(google, 'large_org').infobox('hi').content).toEqual('hi')
-    expect(factory(google, 'large_org').infobox('hi').pixelOffset.x).toEqual(-151)
-    expect(factory(google, 'large_org').infobox('hi').pixelOffset.y).toEqual(-102)
-  });
+    it('Large org infobox', function() {
+      expect(factory(google, 'large_org').infobox('hi').content).toEqual('hi')
+      expect(factory(google, 'large_org').infobox('hi').pixelOffset.x).toEqual(-151)
+      expect(factory(google, 'large_org').infobox('hi').pixelOffset.y).toEqual(10)
+    });
 });
