@@ -3,7 +3,7 @@ require 'rails_helper'
 describe 'VolunteerOps', :type => :request, :helpers => :requests do
   let(:org_owner) { FactoryGirl.create(:user_stubbed_organisation) }
   let(:non_org_owner) { FactoryGirl.create :user , :email => 'regularjoe@blah.com'}
-  let(:admin) {FactoryGirl.create :user, :email => "admin@admin.com", :admin => true}
+  let(:superadmin) {FactoryGirl.create :user, :email => "superadmin@superadmin.com", :superadmin => true}
 
   describe 'POST /volunteer_ops' do
     let(:params) { { volunteer_op: {title: 'hard work', description: 'for the willing'} } }
@@ -31,8 +31,8 @@ describe 'VolunteerOps', :type => :request, :helpers => :requests do
       }.to change(VolunteerOp, :count).by(0)
     end
 
-    it 'does work for admins' do
-      login(admin)
+    it 'does work for superadmins' do
+      login(superadmin)
       expect{
         post organisation_volunteer_ops_path(org_owner.organisation), params
       }.to change(VolunteerOp, :count).by(1)
