@@ -117,20 +117,6 @@ describe 'layouts/application.html.erb', :type => :view do
       expect(rendered).to have_selector('div.alert-error')
     end
 
-    it 'should display a logo linked to the contributors page' do
-      render
-      rendered.within("a[href='#{contributors_path}']") do |hyperlink|
-        expect(hyperlink).to have_css "img[@alt='Agile Ventures']"
-      end
-    end
-
-    it 'should display a logo linked to the ninefold page' do
-      render
-      rendered.within("a[href='https://ninefold.com']") do |hyperlink|
-        expect(hyperlink).to have_css "img[@alt='Ninefold']"
-      end
-    end
-
     it "does not render a new organisation link" do
       allow(view).to receive_messages(:user_signed_in? => false)
       render
@@ -164,7 +150,7 @@ describe 'layouts/application.html.erb', :type => :view do
 
   context 'regular user signed-in' do
     before do
-      allow(user).to receive(:admin?) { false }
+      allow(user).to receive(:superadmin?) { false }
       allow(view).to receive(:current_user) { user }
       render
     end
@@ -178,8 +164,8 @@ describe 'layouts/application.html.erb', :type => :view do
         have_css("li.dropdown ul.dropdown-menu li a[href=\"#{destroy_user_session_path}\"]")
     end
 
-    it 'should not see admin-only dropdown' do
-      expect(rendered).not_to have_css('#menuAdmin')
+    it 'should not see superadmin-only dropdown' do
+      expect(rendered).not_to have_css('#menuSuperAdmin')
     end
 
     it 'does not render a new organisation link' do
@@ -187,9 +173,9 @@ describe 'layouts/application.html.erb', :type => :view do
     end
   end
 
-  context 'admin signed-in' do
+  context 'superadmin signed-in' do
     before(:each) do
-      allow(user).to receive(:admin?) { true }
+      allow(user).to receive(:superadmin?) { true }
       allow(view).to receive(:current_user) { user }
       render
     end
@@ -198,8 +184,8 @@ describe 'layouts/application.html.erb', :type => :view do
       expect(rendered).to have_link('New Organisation', href: new_organisation_path)
     end
 
-    it 'should see admin-only dropdown' do
-      rendered.within('#menuAdmin') do |menu|
+    it 'should see superadmin-only dropdown' do
+      rendered.within('#menuSuperAdmin') do |menu|
         expect(menu).to have_link 'Organisations Without Users', :href => organisations_report_path
         expect(menu).to have_link 'All Users', :href => users_report_path
         expect(menu).to have_link 'Invited Users', :href => invited_users_report_path
