@@ -3,11 +3,11 @@ require 'rails_helper'
 describe "organisations/index.html.erb", :type => :view, :js => true do
 
   let(:org1) do
-    stub_model Organisation,:name => 'test', :address => "12 pinner rd", :postcode => "HA1 4HP",:telephone => "1234", :website => 'http://a.com', :description => 'I am test organisation hahahahahhahaha'
+    stub_model Organisation,:name => 'test', :address => "12 pinner rd", :postcode => "HA1 4HZ",:telephone => "1234", :website => 'http://a.com', :description => 'I am test organisation hahahahahhahaha'
   end
 
   let(:org2) do
-    stub_model Organisation,:name => 'test2', :address => "12 oxford rd", :postcode => "HA1 4HX", :telephone => "4534", :website => 'http://b.com', :description => 'I am '
+    stub_model Organisation,:name => 'test2', :address => "12 oxford rd", :postcode => "HA1 4HZ", :telephone => "4534", :website => 'http://b.com', :description => 'I am '
   end
 
   let(:organisations) do
@@ -21,23 +21,24 @@ describe "organisations/index.html.erb", :type => :view, :js => true do
   before(:each) do
     assign(:organisations, organisations)
     assign(:results, results)
-    assign(:query_term,'search')
-    assign(:category_options, [['Animal Welfare','1'],['Education','2']])
+    assign(:parsed_params, double("ParsedParams", :query_term => 'search'))
+    assign(:cat_name_ids, {what: [['Animal Welfare','1'],['Education','2']],
+      who: [['Youth','3'],['Ethnic Minorities','4']], how: [['Umbrella','5'],['Individual Grants','6']]})
     assign(:markers, 'my-markers')
     render
   end
 
-  it "renders a search form" do
+  it "renders a search form for delivering subcategory-specific params to the controller" do
     expect(rendered).to have_selector "form input[name='q']"
     expect(rendered).to have_selector "form input[type='submit']"
     expect(rendered).to have_selector "form input[value='search']"
-    expect(rendered).to have_selector "form input[placeholder='optional search name/description']"
-    expect(rendered).to have_selector "form select[name='category[id]']"
-    expect(rendered).to have_selector "form select[name='category[id]'] option[value='']" do |all_select|
+    expect(rendered).to have_content "Optional Search Text"
+    expect(rendered).to have_selector "form select[name='what_id']"
+    expect(rendered).to have_selector "form select[name='what_id'] option[value='']" do |all_select|
       expect(all_select).to contain("All")
     end
-    expect(rendered).to have_selector "form select[name='category[id]'] option[value='1']"
-    expect(rendered).to have_selector "form select[name='category[id]'] option[value='2']"
+    expect(rendered).to have_selector "form select[name='what_id'] option[value='1']"
+    expect(rendered).to have_selector "form select[name='what_id'] option[value='2']"
   end
 
   it "render organisation names with hyperlinks" do
@@ -63,7 +64,7 @@ describe "organisations/index.html.erb", :type => :view, :js => true do
 
   it 'displays the json for the map script' do
     assign(:footer_page_links, [])
-    render template: "organisations/index", layout: "layouts/two_columns"
+    render template: "organisations/index", layout: "layouts/two_columns_with_map"
     expect(rendered).to include 'my-markers'
   end
 
