@@ -7,7 +7,7 @@ describe 'organisations/show.html.erb', :type => :view do
         :name => 'Friendly',
         :address => '12 pinner rd',
         :telephone => '1234',
-        :email => 'admin@friendly.org',
+        :email => 'superadmin@friendly.org',
         :postcode => 'HA1 4HZ',
         :website => 'http://www.friendly.org',
         :donation_info => 'http://www.friendly.org/donate',
@@ -20,9 +20,9 @@ describe 'organisations/show.html.erb', :type => :view do
   before(:each) { assign(:organisation, organisation) }
 
   context 'page styling' do
-    it 'name should be wrapped in h3 tag' do
+    it 'name should be wrapped in h2 tag' do
       render
-      expect(rendered).to have_css('h3', :text => organisation.name)
+      expect(rendered).to have_css('h2', :text => organisation.name)
     end
     it 'PRESENT: postcode, email, website, donation info' do
       render
@@ -63,7 +63,7 @@ describe 'organisations/show.html.erb', :type => :view do
           :address => '12 pinner rd',
           :description => 'lovely',
           :telephone => '1234',
-          :email => 'admin@friendly.org',
+          :email => 'superadmin@friendly.org',
           :postcode => 'HA1 4HZ',
           :website => 'http://www.friendly.org',
           :donation_info => 'http://www.friendly.org/donate',
@@ -180,9 +180,9 @@ describe 'organisations/show.html.erb', :type => :view do
     end
   end
 
-  describe 'pending admin status' do
-    it 'displays pending admin message' do
-      assign(:pending_admin, true)
+  describe 'pending superadmin status' do
+    it 'displays pending superadmin message' do
+      assign(:pending_org_admin, true)
       render
       expect(rendered).to have_content 'Your request for admin status is pending.'
     end
@@ -231,8 +231,11 @@ describe 'organisations/show.html.erb', :type => :view do
     end
 
     it 'renders categories when present' do
-      cats = [mock_model(Category, name: "Animal Welfare"), mock_model(Category, name: "Sports")]
+      cats = double(Category::ActiveRecord_AssociationRelation)
       org = mock_model(Organisation, categories: cats)
+      allow(cats).to receive_message_chain(:what_they_do, :pluck).and_return(["Animal Welfare", "Sports"])
+      allow(cats).to receive_message_chain(:who_they_help, :pluck).and_return(["Sports","Sports"])
+      allow(cats).to receive_message_chain(:how_they_help, :pluck).and_return(["Animal Welfare""Sports"])
       assign(:organisation, org)
       render
       expect(rendered).to have_content "Categories:"
@@ -241,5 +244,3 @@ describe 'organisations/show.html.erb', :type => :view do
     end
   end
 end
-
-
