@@ -40,7 +40,7 @@ describe "volunteer_ops/index", :type => :view, :js => true  do
     orgs = [@org1, @org2]
     assign(:footer_page_links, [])
     assign(:markers, 'my-markers')
-    render template: "volunteer_ops/index", layout: "layouts/two_columns"
+    render template: "volunteer_ops/index", layout: "layouts/two_columns_with_map"
     expect(rendered).to include 'my-markers'
   end
 
